@@ -3,65 +3,52 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/date-fns": "^2.16.0",
     "@emotion/babel-preset-css-prop": "^11.10.0",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@fortawesome/fontawesome-svg-core": "^6.4.0",
-    "@fortawesome/free-regular-svg-icons": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@fontsource-variable/roboto-flex": "^5.0.3",
     "@iarna/toml": "^2.2.5",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/pickers": "^3.3.10",
-    "@mui/material": "^5.11.15",
-    "@mui/styles": "^5.11.13",
-    "@mui/x-date-pickers": "^5.0.20",
-    "@reduxjs/toolkit": "^1.8.5",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^14.4.3",
+    "@mui/material": "^5.14.7",
+    "@opencast/appkit": "^0.2.0",
+    "@reduxjs/toolkit": "^1.9.5",
+    "@testing-library/jest-dom": "^6.1.2",
     "@types/iarna__toml": "^2.0.2",
-    "@types/jest": "^29.5.0",
-    "@types/node": "^18.15.11",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^18.0.3",
-    "@types/yup": "^0.32.0",
-    "buffer": "^6.0.3",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
     "customize-cra": "^1.0.0",
     "deepmerge": "^4.3.1",
-    "emotion": "^11.0.0",
     "emotion-normalize": "^11.0.1",
-    "final-form": "^4.20.9",
-    "i18next": "^22.0.4",
-    "i18next-browser-languagedetector": "^7.0.1",
-    "lodash.clonedeep": "^4.5.0",
-    "luxon": "^3.3.0",
-    "mui-rff": "^5.3.2",
-    "react": "^17.0.2",
+    "eslint-plugin-react": "^7.33.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "final-form": "^4.20.10",
+    "i18next": "^23.4.6",
+    "i18next-browser-languagedetector": "^7.1.0",
+    "lodash": "^4.17.21",
+    "luxon": "^3.4.2",
+    "mui-rff": "^6.1.4",
+    "package-lock-only": "^0.0.4",
+    "react": "^18.2.0",
     "react-app-rewired": "^2.2.1",
-    "react-beforeunload": "^2.5.3",
+    "react-beforeunload": "^2.6.0",
     "react-device-detect": "^2.2.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-draggable": "^4.4.5",
     "react-final-form": "^6.5.9",
-    "react-final-form-listeners": "^1.0.3",
     "react-hotkeys": "^2.0.0",
-    "react-i18next": "^12.2.0",
-    "react-indiana-drag-scroll": "^2.1.0",
+    "react-i18next": "^13.2.1",
+    "react-icons": "^4.9.0",
+    "react-indiana-drag-scroll": "^2.2.0",
     "react-player": "git+https://arnei@github.com/Arnei/react-player.git#20fe6c061cf7d71d33d764b4a51c9b9bbb614bf6",
-    "react-redux": "^8.0.5",
-    "react-resizable": "^3.0.4",
+    "react-redux": "^8.1.2",
+    "react-resizable": "^3.0.5",
     "react-scripts": "5.0.1",
-    "react-select": "^5.7.0",
-    "react-use": "^17.4.0",
-    "react-virtualized-auto-sizer": "^1.0.6",
-    "react-window": "^1.8.8",
+    "react-select": "^5.7.4",
+    "react-virtualized-auto-sizer": "^1.0.20",
+    "react-window": "^1.8.9",
     "redux": "^4.2.1",
-    "standardized-audio-context": "^25.3.41",
-    "stream": "0.0.2",
-    "subtitle": "^4.1.2",
+    "standardized-audio-context": "^25.3.55",
     "typescript": "^4.9.5",
+    "uuid": "^9.0.0",
     "webvtt-parser": "^2.2.0"
   },
   "scripts": {
