@@ -8,7 +8,7 @@ import { selectWorkflows, setSelectedWorkflowIndex } from '../redux/videoSlice'
 import { selectFinishState, selectPageNumber } from '../redux/finishSlice'
 
 import { PageButton } from './Finish'
-import { faChevronLeft } from "@fortawesome/free-solid-svg-icons";
+import { LuChevronLeft } from "react-icons/lu";
 import { SaveAndProcessButton } from "./WorkflowConfiguration";
 import { selectStatus, selectError } from "../redux/workflowPostAndProcessSlice";
 import { selectStatus as saveSelectStatus, selectError as saveSelectError } from "../redux/workflowPostSlice";
@@ -16,16 +16,15 @@ import { httpRequestState, Workflow } from "../types";
 import { SaveButton } from "./Save";
 import { EmotionJSX } from "@emotion/react/types/jsx-namespace";
 
-import './../i18n/config';
 import { useTranslation } from 'react-i18next';
 import { Trans } from "react-i18next";
-import { FormControlLabel, Radio, RadioGroup, withStyles } from "@material-ui/core";
-import { selectTheme } from "../redux/themeSlice";
+import { FormControlLabel, Radio, RadioGroup } from "@mui/material";
+import { useTheme } from "../themes";
 
 /**
  * Allows the user to select a workflow
  */
-const WorkflowSelection : React.FC<{}> = () => {
+const WorkflowSelection : React.FC = () => {
 
   const { t } = useTranslation();
 
@@ -35,7 +34,7 @@ const WorkflowSelection : React.FC<{}> = () => {
   const workflows = useSelector(selectWorkflows)
   const finishState = useSelector(selectFinishState)
   const pageNumber = useSelector(selectPageNumber)
-  const theme = useSelector(selectTheme)
+  const theme = useTheme()
 
   const postAndProcessWorkflowStatus = useSelector(selectStatus);
   const postAndProcessError = useSelector(selectError)
@@ -45,7 +44,7 @@ const WorkflowSelection : React.FC<{}> = () => {
   const workflowSelectionStyle = css({
     padding: '20px',
     display: (finishState === "Start processing" && pageNumber === 1) ? 'flex' : 'none',
-    flexDirection: 'column' as const,
+    flexDirection: 'column',
     justifyContent: 'center',
     alignItems: 'center',
     ...(flexGapReplacementStyle(30, false)),
@@ -53,7 +52,7 @@ const WorkflowSelection : React.FC<{}> = () => {
 
   const workflowSelectionSelectionStyle = css({
     display: 'flex',
-    flexDirection: 'column' as const,
+    flexDirection: 'column',
     alignItems: 'left',
     ...(flexGapReplacementStyle(20, false)),
     flexWrap: 'wrap',
@@ -71,7 +70,7 @@ const WorkflowSelection : React.FC<{}> = () => {
   };
 
   // Layout template
-  const render = (topTitle: string, topText: {} | null | undefined, hasWorkflowButtons: boolean,
+  const render = (topTitle: string, topText: JSX.Element, hasWorkflowButtons: boolean,
     nextButton: EmotionJSX.Element, errorStatus: httpRequestState["status"],
     errorMessage: httpRequestState["error"]) => {
     return (
@@ -85,19 +84,19 @@ const WorkflowSelection : React.FC<{}> = () => {
               name="Workflow Selection Area"
               onChange={handleWorkflowSelectChange}
             >
-              {workflows.map( (workflow: Workflow, index: number) => (
+              {workflows.map((workflow: Workflow, _index: number) => (
                 <WorkflowButton key={workflow.id} stateName={workflow.name} workflowId={workflow.id} workflowDescription={workflow.description}/>
               ))}
             </RadioGroup>
         }
         <div css={backOrContinueStyle}>
-          <PageButton pageNumber={0} label={t("workflowSelection.back-button")} iconName={faChevronLeft}/>
+          <PageButton pageNumber={0} label={t("workflowSelection.back-button")} Icon={LuChevronLeft}/>
           {/* <PageButton pageNumber={2} label="Continue" iconName={faChevronRight}/> */}
           {nextButton}
         </div>
         <div css={errorBoxStyle(errorStatus === "failed", theme)} role="alert">
           <span>{t("various.error-text")}</span><br />
-          {errorMessage ? t("various.error-details-text", {errorMessage: postAndProcessError}) : t("various.error-noDetails-text")}<br/>
+          {errorMessage ? t("various.error-details-text", {errorMessage: postAndProcessError}) : t("various.error-text")}<br/>
         </div>
       </div>
     );
@@ -106,12 +105,12 @@ const WorkflowSelection : React.FC<{}> = () => {
   // Fills the layout template with values based on how many workflows are available
   const renderSelection = () => {
     if (workflows.length <= 0) {
-      return(
+      return (
         render(
           t("workflowSelection.saveAndProcess-text"),
           <Trans i18nKey="workflowSelection.noWorkflows-text">
-            A problem occurred, there are no workflows to process your changes with.<br />
-            Please save your changes and contact an Opencast Administrator.
+            There are no workflows to process your changes with.<br />
+            Please save your changes and contact an administrator.
           </Trans>,
           false,
           <SaveButton />,
@@ -124,7 +123,7 @@ const WorkflowSelection : React.FC<{}> = () => {
         render(
           t("workflowSelection.saveAndProcess-text"),
           <Trans i18nKey="workflowSelection.oneWorkflow-text">
-            The video will be cut and processed with the workflow "{{workflow: workflows[0].name}}".<br/>
+            The video will be cut and processed with the workflow {{workflow: workflows[0].name}}.<br/>
             This will take some time.
           </Trans>,
           false,
@@ -161,12 +160,12 @@ const WorkflowButton: React.FC<{stateName: string, workflowId: string, workflowD
   const labelStyle = css({
     display: 'flex',
     flexDirection: 'column',
-    maxWidth: '500px'
+    maxWidth: '500px',
   })
 
   const headerStyle = css({
     width: '100%',
-    padding: '5px',
+    padding: '5px 0px',
     fontSize: 'larger',
   });
 
@@ -182,15 +181,26 @@ const WorkflowButton: React.FC<{stateName: string, workflowId: string, workflowD
   );
 }
 
-const WorkflowSelectRadio = withStyles({
-  root: {
+const WorkflowSelectRadio: React.FC = props => {
+
+  const theme = useTheme()
+
+  const style = css({
     alignSelf: 'start',
-    color: 'grey',
+    color: `${theme.text}`,
     "&$checked": {
-      color: 'grey'
+      color: `${theme.text}`
     }
-  },
-  checked: {}
-})((props) => <Radio color="default" {...props} />);
+  })
+
+  return (
+    <Radio
+      color="default"
+      css={style}
+      {...props}
+    />
+  )
+}
+
 
 export default WorkflowSelection;
