@@ -1,7 +1,6 @@
+// Overrides to make emotion css work with createReactApp
+const { override, addBabelPreset } = require('customize-cra')
 
-  // Overrides to make emotion css work with createReactApp
-  const { override, addBabelPreset } = require('customize-cra')
-  
-  module.exports = override(
-    addBabelPreset('@emotion/babel-preset-css-prop')
-  )
+module.exports = override(
+  addBabelPreset('@emotion/babel-preset-css-prop')
+)
