@@ -2,8 +2,10 @@ import { PlaywrightTestConfig, devices } from '@playwright/test';
 
 const config: PlaywrightTestConfig = {
 
-  reporter: [ [process.env.CI ? 'github' : 'list',],
-            ['html', { outputFolder: 'playwright-report' }]],
+  reporter: [
+    [process.env.CI ? 'github' : 'list'],
+    ['html', { outputFolder: 'playwright-report' }],
+  ],
   testIgnore: '**/redux/**',
   retries: 1,
   timeout: 60 * 1000,
