@@ -11,6 +11,7 @@ for (const lang of locales) {
   const code = lang.replace(/\..*$/, '');
   const short = code.replace(/-.*$/, '');
   const main = locales.filter(l => l.indexOf(short) === 0).length === 1
+  /* eslint-disable-next-line @typescript-eslint/no-var-requires */
   const translations = require('./locales/' + lang);
   if (!main) {
     resources[code] = { translation: translations };
@@ -19,13 +20,13 @@ for (const lang of locales) {
 }
 
 i18next
-    .use(initReactI18next)
-    .use(LanguageDetector)
-    .init({
-        resources,
-        fallbackLng: ['en-US', 'en'],
-        nonExplicitSupportedLngs: true,
-        debug: debug,
+  .use(initReactI18next)
+  .use(LanguageDetector)
+  .init({
+    resources,
+    fallbackLng: ['en-US', 'en'],
+    nonExplicitSupportedLngs: true,
+    debug: debug,
   });
 
 if (debug) {
