@@ -8,10 +8,14 @@ import store from './redux/store'
 import { init } from './config'
 import { sleep } from './util/utilityFunctions'
 
-import { MuiPickersUtilsProvider } from '@material-ui/pickers';
-import DateFnsUtils from '@date-io/date-fns';
 import { GlobalHotKeys } from 'react-hotkeys';
 
+import "@fontsource-variable/roboto-flex";
+
+import './i18n/config';
+
+import '@opencast/appkit/dist/colors.css'
+import { ColorSchemeProvider } from '@opencast/appkit';
 
 // Load config here
 // Load the rest of the application and try to fetch the settings file from the
@@ -30,14 +34,14 @@ initialize.then(
   () => {
     ReactDOM.render(
       <React.StrictMode>
-          <Provider store={store}>
-            <MuiPickersUtilsProvider utils={DateFnsUtils}>
-              {/* Workaround for getApplicationKeyMap based on https://github.com/greena13/react-hotkeys/issues/228 */}
-              <GlobalHotKeys>
-                <App />
-              </GlobalHotKeys>
-            </MuiPickersUtilsProvider>
-          </Provider>
+        <Provider store={store}>
+          {/* Workaround for getApplicationKeyMap based on https://github.com/greena13/react-hotkeys/issues/228 */}
+          <GlobalHotKeys>
+            <ColorSchemeProvider>
+              <App />
+            </ColorSchemeProvider>
+          </GlobalHotKeys>
+        </Provider>
       </React.StrictMode>,
       document.getElementById('root')
     );
