@@ -6,35 +6,36 @@ import Discard from "./Discard"
 import WorkflowSelection from "./WorkflowSelection";
 import WorkflowConfiguration from "./WorkflowConfiguration";
 
-import { css } from '@emotion/react'
-import { basicButtonStyle } from '../cssStyles'
+import { LuDoorOpen} from "react-icons/lu";
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  IconDefinition
-} from "@fortawesome/free-solid-svg-icons";
+import { css } from '@emotion/react'
+import { basicButtonStyle, navigationButtonStyle } from '../cssStyles'
+
+import { IconType } from "react-icons";
 
 import { useDispatch, useSelector } from 'react-redux';
 import { selectPageNumber, setPageNumber } from '../redux/finishSlice';
-import { selectTheme } from "../redux/themeSlice";
+import { useTheme } from "../themes";
+import { settings } from "../config";
+import { useTranslation } from "react-i18next";
 
 /**
  * Displays a menu for selecting what should be done with the current changes
  */
-const Finish : React.FC<{}> = () => {
+const Finish : React.FC = () => {
 
   const pageNumber = useSelector(selectPageNumber)
 
   const pageZeroStyle = css({
-    display: pageNumber !== 0 ? 'none' :'block',
+    display: pageNumber !== 0 ? 'none' : 'block',
   })
 
   const pageOneStyle = css({
-    display: pageNumber !== 1 ? 'none' :'block',
+    display: pageNumber !== 1 ? 'none' : 'block',
   })
 
   const pageTwoStyle = css({
-    display: pageNumber !== 2 ? 'none' :'block',
+    display: pageNumber !== 2 ? 'none' : 'block',
   })
 
   return (
@@ -57,10 +58,10 @@ const Finish : React.FC<{}> = () => {
 /**
  * Takes you to a different page
  */
-export const PageButton : React.FC<{pageNumber: number, label: string, iconName: IconDefinition}> = ({pageNumber, label, iconName}) => {
+export const PageButton : React.FC<{pageNumber: number, label: string, Icon: IconType}> = ({pageNumber, label, Icon}) => {
 
-  const theme = useSelector(selectTheme);
-  
+  const theme = useTheme();
+
   // Initialize redux variables
   const dispatch = useDispatch()
 
@@ -69,9 +70,9 @@ export const PageButton : React.FC<{pageNumber: number, label: string, iconName:
   }
 
   const pageButtonStyle = css({
-    width: '200px',
+    minWidth: '100px',
     padding: '16px',
-    justifyContent: 'space-around',
+    justifyContent: 'center',
     boxShadow: `${theme.boxShadow}`,
     background: `${theme.element_bg}`,
   })
@@ -79,13 +80,43 @@ export const PageButton : React.FC<{pageNumber: number, label: string, iconName:
   return (
     <div css={[basicButtonStyle(theme), pageButtonStyle]}
       role="button" tabIndex={0}
-      onClick={ onPageChange }
+      onClick={onPageChange}
       onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
         onPageChange()
-      }}}>
-      <FontAwesomeIcon icon={iconName} size="1x" />
+      } }}>
+      <Icon />
       <span>{label}</span>
     </div>
+  );
+}
+
+/**
+ * Takes you back to the callback url resource
+ */
+export const CallbackButton : React.FC = () => {
+
+  const { t } = useTranslation();
+
+  const theme = useTheme();
+
+  const openCallbackUrl = () => {
+    window.open(settings.callbackUrl, "_self");
+  }
+
+  return (
+    <>
+      {settings.callbackUrl !== undefined &&
+        <div css={[basicButtonStyle(theme), navigationButtonStyle(theme)]}
+          role="button" tabIndex={0}
+          onClick={openCallbackUrl}
+          onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
+            openCallbackUrl()
+          } }}>
+          <LuDoorOpen />
+          <span>{settings.callbackSystem ? t("various.callback-button-system", {system: settings.callbackSystem}) : t("various.callback-button-generic")}</span>
+        </div>
+      }
+    </>
   );
 }
 
