@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 
 import { css } from '@emotion/react'
-import { calendarStyle, errorBoxStyle, selectFieldStyle } from '../cssStyles'
+import { calendarStyle, errorBoxStyle, selectFieldStyle, titleStyle, titleStyleBold } from '../cssStyles'
 
 import { useSelector, useDispatch } from 'react-redux';
 import {
@@ -13,24 +13,15 @@ import { Form, Field, FieldInputProps } from 'react-final-form'
 import Select from 'react-select'
 import CreatableSelect from 'react-select/creatable';
 
-import {
-  DateTimePicker,
-  TimePicker,
-  showErrorOnBlur,
-} from 'mui-rff';
-import DateFnsUtils from "@date-io/date-fns";
-
-import './../i18n/config';
-import i18next from "./../i18n/config";
 import { useTranslation } from 'react-i18next';
 import { DateTime as LuxonDateTime} from "luxon";
 
 import { configureFieldsAttributes, settings } from '../config'
-import { LocalizationProvider } from "@mui/x-date-pickers";
-import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { AppDispatch } from "../redux/store";
-import { selectTheme } from "../redux/themeSlice";
+import { useTheme } from "../themes";
 import { ThemeProvider } from "@mui/material/styles";
+import { cloneDeep } from "lodash";
+import { ParseKeys } from "i18next";
 
 /**
  * Creates a Metadata form
@@ -41,7 +32,7 @@ import { ThemeProvider } from "@mui/material/styles";
  * If something doesn't work, main places of interest are the submit function
  * and the initialValues function
  */
-const Metadata: React.FC<{}> = () => {
+const Metadata: React.FC = () => {
 
   const { t, i18n } = useTranslation();
 
@@ -52,7 +43,7 @@ const Metadata: React.FC<{}> = () => {
   const getError = useSelector(selectGetError);
   const postStatus = useSelector(selectPostStatus);
   const postError = useSelector(selectPostError);
-  const theme = useSelector(selectTheme);
+  const theme = useTheme();
 
   // Try to fetch URL from external API
   useEffect(() => {
@@ -64,14 +55,14 @@ const Metadata: React.FC<{}> = () => {
   // Overwrite readonly property of fields based on config settings
   useEffect(() => {
     if (getStatus === 'success') {
-      for(let catalogIndex = 0; catalogIndex < catalogs.length; catalogIndex++) {
+      for (let catalogIndex = 0; catalogIndex < catalogs.length; catalogIndex++) {
         if (settings.metadata.configureFields) {
-          let configureFields = settings.metadata.configureFields
-          let catalog = catalogs[catalogIndex]
+          const configureFields = settings.metadata.configureFields
+          const catalog = catalogs[catalogIndex]
 
           if (catalog.title in configureFields) {
             if (Object.keys(configureFields[catalog.title]).length > 0) {
-              let configureFieldsCatalog = configureFields[catalog.title]
+              const configureFieldsCatalog = configureFields[catalog.title]
 
               for (let fieldIndex = 0; fieldIndex < catalog.fields.length; fieldIndex++) {
                 if (catalog.fields[fieldIndex].id in configureFieldsCatalog) {
@@ -97,36 +88,50 @@ const Metadata: React.FC<{}> = () => {
 
   const metadataStyle = css({
     padding: '20px',
-    marginLeft:'auto',
-    marginRight:'auto',
+    marginLeft: 'auto',
+    marginRight: 'auto',
     minWidth: '50%',
     display: 'grid',
   })
 
+  const catalogStyle = css({
+    background: `${theme.menu_background}`,
+    borderRadius: '5px',
+    boxShadow: `${theme.boxShadow_tiles}`,
+    marginTop: '24px',
+    boxSizing: "border-box",
+    padding: '10px',
+  })
+
   const fieldStyle = css({
     display: 'flex',
-    flexFlow: 'row nowrap',
+    flexFlow: 'column nowrap',
     lineHeight: '2em',
     margin: '10px',
   })
 
   const fieldLabelStyle = css({
-    alignSelf: 'center',
     width: '110px',
     fontSize: '1em',
+    fontWeight: 'bold',
+    color: `${theme.text}`,
     lineHeight: '32px',
   })
 
   const fieldTypeStyle = (isReadOnly: boolean) => {
     return css({
-      flex: '1',
       fontSize: '1em',
-      marginLeft: '15px',
       borderRadius: '5px',
       boxShadow: isReadOnly ? '0 0 0px rgba(0, 0, 0, 0.3)' : '0 0 1px rgba(0, 0, 0, 0.3)',
       ...(isReadOnly && {color: `${theme.text}`}),
       color: `${theme.text}`,
-      outline: isReadOnly ? '0px solid transparent' : `${theme.element_outline}`
+      outline: isReadOnly ? '0px solid transparent' : `${theme.element_outline}`,
+      "&:hover": {
+        borderColor: isReadOnly ? undefined : theme.metadata_highlight,
+      },
+      "&:focus": {
+        borderColor: isReadOnly ? undefined : theme.metadata_highlight,
+      },
     });
   }
 
@@ -134,33 +139,10 @@ const Metadata: React.FC<{}> = () => {
     return (
       css({
         padding: '10px 10px',
-        border: isReadOnly ? '0px solid #ccc' : '1px solid #ccc',
+        border: '1px solid #ccc',
         background: isReadOnly ? `${theme.background}` : `${theme.element_bg}`,
-      })
-    );
-  }
-
-  const dateTimeTypeStyle = (isReadOnly: boolean) => {
-    return (
-      css ({
-        padding: '5px 10px',
-        border: isReadOnly ? '0px solid #ccc' : '1px solid #ccc',
-        background: isReadOnly ? `${theme.background}` : `${theme.element_bg}`,
-        '.Mui-disabled, .Mui-disabled button > svg': {
-          color: `${theme.disabled} !important`,
-          'WebkitTextFillColor':`${theme.disabled}`,
-        },
-        'button > svg': {
-          color: `${theme.indicator_color}`
-        },
-        '.MuiInput-input, button': {
-          color: `${theme.text}`,
-          background: 'transparent !important',
-          '&:hover': {
-            background: 'transparent !important',
-            outline: 'none'
-          }
-        },    
+        opacity: isReadOnly ? "0.6" : "1",
+        resize: 'vertical',
       })
     );
   }
@@ -219,7 +201,7 @@ const Metadata: React.FC<{}> = () => {
    * Form Callbacks - Other
    */
 
-   /**
+  /**
     * Recursively recreates nested array structures for form initalValues
     * @param library
     * @param input
@@ -246,7 +228,7 @@ const Metadata: React.FC<{}> = () => {
 
     catalogs.forEach((catalog: Catalog, catalogIndex: number) => {
       initValues["catalog" + catalogIndex] = {}
-      catalog.fields.forEach((field: MetadataField) =>{
+      catalog.fields.forEach((field: MetadataField) => {
         initValues["catalog" + catalogIndex][field.id] = field.value
 
         // Handle initial values for select fields differently
@@ -256,7 +238,7 @@ const Metadata: React.FC<{}> = () => {
           let searchValue : any = field.value
 
           if (Array.isArray(searchValue)) {
-            let result: any[] = [];
+            const result: any[] = [];
             helperHandleArrays(library, field.value, result)
             searchValue = result
           } else {
@@ -286,7 +268,7 @@ const Metadata: React.FC<{}> = () => {
    * @param value
    */
   const duration = (value: any) => {
-    let re: RegExp = /^[0-9][0-9]:[0-9][0-9]:[0-9][0-9]$/
+    const re = /^[0-9][0-9]:[0-9][0-9]:[0-9][0-9]$/
     return re.test(value) ? undefined : t("metadata.validation.duration-format")
   }
 
@@ -304,7 +286,7 @@ const Metadata: React.FC<{}> = () => {
     if (Object.prototype.toString.call(date) === '[object Date]') {
       dt = LuxonDateTime.fromJSDate(date);
     }
-    if (typeof(date) === 'string') {
+    if (typeof date === 'string') {
       dt = LuxonDateTime.fromISO(date);
     }
 
@@ -339,7 +321,7 @@ const Metadata: React.FC<{}> = () => {
    * Form Callbacks - Submitting
    */
 
-   /**
+  /**
     * Sends a single value to the corresponding field in redux.
     * This kinda breaks the form workflow, since we do not use the submit callback
     * of the form class anymore.
@@ -362,7 +344,8 @@ const Metadata: React.FC<{}> = () => {
       if (catalogs[catalogIndex].fields[fieldIndex].id === fieldName) {
         // Update the field in the redux catalog
         dispatch(setFieldValue({catalogIndex: catalogIndex, fieldIndex: fieldIndex,
-          value: parseValue(catalogs[catalogIndex].fields[fieldIndex], value)}))
+          value: parseValue(catalogs[catalogIndex].fields[fieldIndex], value)
+        }))
         break
       }
     }
@@ -374,8 +357,8 @@ const Metadata: React.FC<{}> = () => {
    * @param input
    */
   const blurWithSubmit = (e: any, input: any) => {
-      input.onBlur(e);
-      submitSingleField(input.value, input.name)
+    input.onBlur(e);
+    submitSingleField(input.value, input.name)
   }
 
   /**
@@ -388,7 +371,7 @@ const Metadata: React.FC<{}> = () => {
     let returnValue : any = value
 
     // Parse values out react-multi-select and put them in an array
-    if(Array.isArray(value)) {
+    if (Array.isArray(value)) {
       returnValue = []
       value.forEach((subValue : any) => {
         returnValue.push(parseValue(null, subValue))  // Pass field as null to avoid each value into an array later on
@@ -396,14 +379,14 @@ const Metadata: React.FC<{}> = () => {
     }
 
     // If the value is hidden an object due to react-select, extract it
-    if (typeof value === 'object' && value !== null && value.hasOwnProperty("submitValue")) {
+    if (typeof value === 'object' && value !== null && Object.prototype.hasOwnProperty.call(value, "submitValue")) {
       returnValue = value.submitValue
     } else if (typeof value === 'object' && value !== null && value.__isNew__) {
       returnValue = value.value
     }
 
     // For these fields, the value needs to be inside an array
-    if (field && !Array.isArray(value) &&(field.id === "creator" || field.id === "contributor")) {
+    if (field && !Array.isArray(value) && (field.id === "creator" || field.id === "contributor")) {
       returnValue = [returnValue]
     }
 
@@ -432,7 +415,7 @@ const Metadata: React.FC<{}> = () => {
   const onSubmit = (values: { [x: string]: { [x: string]: any; }; }) => {
     // For each submitted value, get the catalog it belongs to
     Object.keys(values).forEach((formCatalogName: string) => {
-      let catalogIndex = parseInt(formCatalogName.replace("catalog", ""))
+      const catalogIndex = parseInt(formCatalogName.replace("catalog", ""))
 
       // For each field in the submitted values
       Object.keys(values[formCatalogName]).forEach((formFieldName: any) => {
@@ -441,7 +424,8 @@ const Metadata: React.FC<{}> = () => {
           if (catalogs[catalogIndex].fields[fieldIndex].id === formFieldName) {
             // Update the field in the redux catalog
             dispatch(setFieldValue({catalogIndex: catalogIndex, fieldIndex: fieldIndex,
-              value: parseValue(catalogs[catalogIndex].fields[fieldIndex], values[formCatalogName][formFieldName])}))
+              value: parseValue(catalogs[catalogIndex].fields[fieldIndex], values[formCatalogName][formFieldName])
+            }))
             break
           }
         }
@@ -475,10 +459,10 @@ const Metadata: React.FC<{}> = () => {
         // Parse Label
         let descLabel = null
         if (i18n.exists(`metadata.${field.id}`)) {
-          descLabel = t(`metadata.${field.id}.${key.replaceAll(".", "-")}`)
+          descLabel = t(`metadata.${field.id}.${key.replaceAll(".", "-")}` as ParseKeys)
 
           if (field.id === "license") {
-            descLabel = t(`metadata.${field.id}.${JSON.parse(key).label.replaceAll(".", "-")}`)
+            descLabel = t(`metadata.${field.id}.${JSON.parse(key).label.replaceAll(".", "-")}` as ParseKeys)
           }
         }
 
@@ -511,7 +495,7 @@ const Metadata: React.FC<{}> = () => {
       if (Array.isArray(field.value)) {
         return (
           <CreatableSelect {...input}
-            onBlur={e => {blurWithSubmit(e, input)}}
+            onBlur={e => { blurWithSubmit(e, input) }}
             isMulti
             isClearable={!field.readOnly}     // The component does not support readOnly, so we have to work around
             isSearchable={!field.readOnly}    // by setting other settings
@@ -521,11 +505,11 @@ const Metadata: React.FC<{}> = () => {
             styles={selectFieldStyle(theme)}
             css={fieldTypeStyle(field.readOnly)}>
           </CreatableSelect>
-          );
+        );
       } else {
         return (
           <Select {...input}
-            onBlur={e => {blurWithSubmit(e, input)}}
+            onBlur={e => { blurWithSubmit(e, input) }}
             isClearable={!field.readOnly}     // The component does not support readOnly, so we have to work around
             isSearchable={!field.readOnly}    // by setting other settings
             openMenuOnClick={!field.readOnly}
@@ -534,61 +518,56 @@ const Metadata: React.FC<{}> = () => {
             styles={selectFieldStyle(theme)}
             css={fieldTypeStyle(field.readOnly)}>
           </Select>
-          );
+        );
       }
 
     } else if (field.type === "date") {
       return (
-        <div data-testid="dateTimePicker" css={[fieldTypeStyle(field.readOnly), dateTimeTypeStyle(field.readOnly)]}>
-          <LocalizationProvider dateAdapter={AdapterDateFns}>
-            <ThemeProvider theme={calendarStyle(theme)}>
-              <DateTimePicker {...input}
-                name={field.id}
-                inputFormat="yyyy/MM/dd HH:mm"
-                disabled={field.readOnly}
-                dateFunsUtils={DateFnsUtils}
-                TextFieldProps={{
-                  variant: 'standard', // Removes default outline
-                  onBlur: (e: any) => {blurWithSubmit(e, input)},
-                  showError: showErrorOnBlur
-                }}
-                leftArrowButtonText={i18next.t('metadata.calendar-prev')}
-                rightArrowButtonText={i18next.t('metadata.calendar-next')}
-              />
-            </ThemeProvider>
-          </LocalizationProvider>
-        </div>
+        <ThemeProvider theme={calendarStyle(theme)}>
+          <input {...input}
+            type="datetime-local"
+            name={field.id}
+            // inputFormat="yyyy/MM/dd HH:mm"
+            onBlur={e => { blurWithSubmit(e, input) }}
+            readOnly={field.readOnly}
+            css={[fieldTypeStyle(field.readOnly), inputFieldTypeStyle(field.readOnly),
+              {
+                resize: 'none',
+              }
+            ]}
+            data-testid="dateTimePicker"
+          />
+        </ThemeProvider>
       );
     } else if (field.type === "time") {
       return (
-        <div css={[fieldTypeStyle(field.readOnly), dateTimeTypeStyle(field.readOnly)]}>
-          <ThemeProvider theme={calendarStyle(theme)}>
-            <TimePicker {...input}
-              name={field.id}
-              inputFormat="HH:mm"
-              disabled={field.readOnly}
-              dateFunsUtils={DateFnsUtils}
-              TextFieldProps={{
-                variant: 'standard', // Removes default outline
-                onBlur: (e: any) => {blurWithSubmit(e, input)},
-                showError: showErrorOnBlur
-              }}
-            />
-          </ThemeProvider>
-        </div>
+        <ThemeProvider theme={calendarStyle(theme)}>
+          <input {...input}
+            type="time"
+            name={field.id}
+            // inputFormat="HH:mm"
+            onBlur={e => { blurWithSubmit(e, input) }}
+            readOnly={field.readOnly}
+            css={[fieldTypeStyle(field.readOnly), inputFieldTypeStyle(field.readOnly),
+              {
+                resize: 'none',
+              }
+            ]}
+          />
+        </ThemeProvider>
       );
     } else if (field.type === "text_long") {
       return (
         <textarea {...input}
-          onBlur={e => {blurWithSubmit(e, input)}}
+          onBlur={e => { blurWithSubmit(e, input) }}
           readOnly={field.readOnly}
           css={[fieldTypeStyle(field.readOnly), inputFieldTypeStyle(field.readOnly)]}
         />
       );
     } else {
-      return(
+      return (
         <input {...input}
-          onBlur={e => {blurWithSubmit(e, input)}}
+          onBlur={e => { blurWithSubmit(e, input) }}
           readOnly={field.readOnly}
           css={[fieldTypeStyle(field.readOnly), inputFieldTypeStyle(field.readOnly)]}
         />
@@ -612,31 +591,43 @@ const Metadata: React.FC<{}> = () => {
      */
     const generateComponentWithModifiedInput = (field: MetadataField, input: FieldInputProps<any, HTMLElement>) => {
       if ((field.type === "date" || field.type === "time") && input.value === "") {
-        var {value, ...other} = input
+        const {value, ...other} = input
         return generateComponent(field, other)
+      }
+      // <input type="datetime-local"> is picky about its value and won't accept
+      // global datetime strings, so we have to convert them to local ourselves.
+      // TODO: Also we really should not be modifying the input element like that
+      // so ideally the conversion happens somewhere else in the code
+      // (see error in the console for further details)
+      if ((field.type === "date" || field.type === "time")) {
+        input = cloneDeep(input)
+        const leDate = new Date(input.value)
+        leDate.setMinutes(leDate.getMinutes() - leDate.getTimezoneOffset());
+        input.value = leDate.toISOString().slice(0, 16);
+        return generateComponent(field, input)
       } else {
         return generateComponent(field, input)
       }
     }
 
     return (
-        <Field key={fieldIndex}
-                name={"catalog" + catalogIndex + "." + field.id}
-                validate={getValidators(field)}
-                type={field.type === "boolean" ? "checkbox" : undefined}  // react-final-form complains if we don't specify checkboxes here
-                >
-                {({ input, meta }) => (
-                  <div css={fieldStyle} data-testid={field.id}>
-                    <label css={fieldLabelStyle} htmlFor={input.name}>{
-                      i18n.exists(`metadata.labels.${field.id}`) ?
-                      t(`metadata.labels.${field.id}`) : field.id
-                    }</label>
+      <Field key={fieldIndex}
+        name={"catalog" + catalogIndex + "." + field.id}
+        validate={getValidators(field)}
+        type={field.type === "boolean" ? "checkbox" : undefined}  // react-final-form complains if we don't specify checkboxes here
+      >
+        {({ input, meta }) => (
+          <div css={fieldStyle} data-testid={field.id}>
+            <label css={fieldLabelStyle} htmlFor={input.name}>{
+              i18n.exists(`metadata.labels.${field.id}`) ?
+                t(`metadata.labels.${field.id}` as ParseKeys) as string : field.id
+            }</label>
 
-                    {generateComponentWithModifiedInput(field, input)}
-                    {meta.error && meta.touched && <span css={validateStyle(true)}>{meta.error}</span>}
-                  </div>
-                )}
-        </Field>
+            {generateComponentWithModifiedInput(field, input)}
+            {meta.error && meta.touched && <span css={validateStyle(true)}>{meta.error}</span>}
+          </div>
+        )}
+      </Field>
     );
   }
 
@@ -645,13 +636,15 @@ const Metadata: React.FC<{}> = () => {
     catalogIndex: number,
     configureFields: { [key: string]: configureFieldsAttributes }
   ) => {
+
+
     return (
-      <div key={catalogIndex}>
-        <h2>
+      <div key={catalogIndex} css={catalogStyle}>
+        <div css={[titleStyle(theme), titleStyleBold(theme)]}>
           {i18n.exists(`metadata.${catalog.title.replaceAll(".", "-")}`) ?
-            t(`metadata.${catalog.title.replaceAll(".", "-")}`) : catalog.title
+            t(`metadata.${catalog.title.replaceAll(".", "-")}` as ParseKeys) as string : catalog.title
           }
-        </h2>
+        </div>
 
         {catalog.fields.map((field, i) => {
           // Render fields based on given array (usually parsed from config settings)
@@ -674,38 +667,38 @@ const Metadata: React.FC<{}> = () => {
    */
   const render = () => {
     return (
-        <Form
-          onSubmit={onSubmit}
-          subscription={{ submitting: true, pristine: true }} // Hopefully causes less rerenders
-          initialValues={getInitialValues(catalogs)}
-          render={({ handleSubmit, form, submitting, pristine, values}) => (
-            <form onSubmit={event => {
-              handleSubmit(event)
-              // Ugly fix for form not getting updated after submit. TODO: Find a better fix
-              form.reset()
-            }} css={metadataStyle}>
+      <Form
+        onSubmit={onSubmit}
+        subscription={{ submitting: true, pristine: true }} // Hopefully causes less rerenders
+        initialValues={getInitialValues(catalogs)}
+        render={({ handleSubmit, form }) => (
+          <form onSubmit={event => {
+            handleSubmit(event)
+            // Ugly fix for form not getting updated after submit. TODO: Find a better fix
+            form.reset()
+          }} css={metadataStyle}>
 
-              <div css={errorBoxStyle(getStatus === "failed", theme)} role="alert">
-                <span>A problem occurred during communication with Opencast.</span><br />
-                {getError ? "Details: " + getError : "No error details are available."}<br />
-              </div>
+            <div css={errorBoxStyle(getStatus === "failed", theme)} role="alert">
+              <span>A problem occurred during communication with Opencast.</span><br />
+              {getError ? "Details: " + getError : "No error details are available."}<br />
+            </div>
 
-              {catalogs.map((catalog, i) => {
-                if (settings.metadata.configureFields) {
-                  if (catalog.title in settings.metadata.configureFields) {
-                    // If there are no fields for a given catalog, do not render
-                    if (Object.keys(settings.metadata.configureFields[catalog.title]).length > 0) {
-                      return renderCatalog(catalog, i, settings.metadata.configureFields[catalog.title])
-                    } else {
-                      return undefined
-                    }
+            {catalogs.map((catalog, i) => {
+              if (settings.metadata.configureFields) {
+                if (catalog.title in settings.metadata.configureFields) {
+                  // If there are no fields for a given catalog, do not render
+                  if (Object.keys(settings.metadata.configureFields[catalog.title]).length > 0) {
+                    return renderCatalog(catalog, i, settings.metadata.configureFields[catalog.title])
+                  } else {
+                    return undefined
                   }
                 }
-                // If there are no settings for a given catalog, just render it completely
-                return renderCatalog(catalog, i, {})
-              })}
+              }
+              // If there are no settings for a given catalog, just render it completely
+              return renderCatalog(catalog, i, {})
+            })}
 
-{/*
+            {/*
                 <div css={{display: "block", wordWrap: "normal", whiteSpace: "pre"}}>{t("metadata.submit-helpertext", { buttonName: t("metadata.submit-button") })}</div>
 
 
@@ -727,21 +720,21 @@ const Metadata: React.FC<{}> = () => {
                 </button>
               </div> */}
 
-              <div css={errorBoxStyle(postStatus === "failed", theme)} role="alert">
-                <span>A problem occurred during communication with Opencast. <br />
+            <div css={errorBoxStyle(postStatus === "failed", theme)} role="alert">
+              <span>A problem occurred during communication with Opencast. <br />
                       Changes could not be saved to Opencast.</span><br />
-                {postError ? "Details: " + postError : "No error details are available."}<br />
-              </div>
+              {postError ? "Details: " + postError : "No error details are available."}<br />
+            </div>
 
-              {/* For debugging the forms current values*/}
-              {/* <FormSpy subscription={{ values: true }}>
+            {/* For debugging the forms current values*/}
+            {/* <FormSpy subscription={{ values: true }}>
                 {({ values }) => (
                   <pre>{JSON.stringify(values, null, 2)}</pre>
                 )}
               </FormSpy> */}
-            </form>
-          )}
-        />
+          </form>
+        )}
+      />
     );
   }
 
