@@ -3,8 +3,8 @@ import { css } from "@emotion/react";
 import { SegmentsList as CuttingSegmentsList, Waveforms } from "./Timeline";
 import {
   selectCurrentlyAt,
-  selectSelectedSubtitleByFlavor,
-  selectSelectedSubtitleFlavor,
+  selectSelectedSubtitleById,
+  selectSelectedSubtitleId,
   setClickTriggered,
   setCueAtIndex,
   setCurrentlyAt,
@@ -22,26 +22,27 @@ import "react-resizable/css/styles.css";
 import { GlobalHotKeys } from "react-hotkeys";
 import { scrubberKeyMap } from "../globalKeys";
 import ScrollContainer, { ScrollEvent } from "react-indiana-drag-scroll";
-import { selectTheme } from "../redux/themeSlice";
+import { useTheme } from "../themes";
 import { ThemedTooltip } from "./Tooltip";
-import { t } from "i18next";
+import { useTranslation } from "react-i18next";
 
 /**
  * Copy-paste of the timeline in Video.tsx, so that we can make some small adjustments,
  * like adding in a list of subtitle segments
  */
- const SubtitleTimeline: React.FC<{}> = () => {
+const SubtitleTimeline: React.FC = () => {
 
-  const theme = useSelector(selectTheme)
+  const { t } = useTranslation();
+  const theme = useTheme()
 
   // Init redux variables
   const dispatch = useDispatch();
   const duration = useSelector(selectDuration)
   const currentlyAt = useSelector(selectCurrentlyAt)
 
-  const { ref, width = 1, } = useResizeObserver<HTMLDivElement>();
-  const refTop = useRef<ScrollContainer>(null);
-  const { ref: refMini, width: widthMiniTimeline = 1, } = useResizeObserver<HTMLDivElement>();
+  const { ref, width = 1 } = useResizeObserver<HTMLDivElement>();
+  const refTop = useRef<HTMLElement>(null);
+  const { ref: refMini, width: widthMiniTimeline = 1 } = useResizeObserver<HTMLDivElement>();
 
   const timelineCutoutInMs = 10000    // How much of the timeline should be visible in milliseconds. Aka a specific zoom level
 
@@ -53,8 +54,8 @@ import { t } from "i18next";
   });
 
   const setCurrentlyAtToClick = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    let rect = e.currentTarget.getBoundingClientRect()
-    let offsetX = e.clientX - rect.left
+    const rect = e.currentTarget.getBoundingClientRect()
+    const offsetX = e.clientX - rect.left
     dispatch(setClickTriggered(true))
     dispatch(setCurrentlyAt((offsetX / widthMiniTimeline) * (duration)))
   }
@@ -62,8 +63,8 @@ import { t } from "i18next";
   // Apply horizonal scrolling when scrolled from somewhere else
   useEffect(() => {
     if (currentlyAt !== undefined && refTop.current) {
-      const scrollLeftMax = (refTop.current.getElement().scrollWidth - refTop.current.getElement().clientWidth)
-      refTop.current.getElement().scrollTo(Math.round((currentlyAt / duration) * scrollLeftMax), 0)
+      const scrollLeftMax = (refTop.current.scrollWidth - refTop.current.clientWidth)
+      refTop.current.scrollTo(Math.round((currentlyAt / duration) * scrollLeftMax), 0)
     }
   }, [currentlyAt, duration, width]);
 
@@ -83,8 +84,8 @@ import { t } from "i18next";
   const onEndScroll = (e: ScrollEvent) => {
     // If scrolled by user
     if (!e.external && refTop && refTop.current) {
-      const offsetX = refTop.current.getElement().scrollLeft
-      const scrollLeftMax = (refTop.current.getElement().scrollWidth - refTop.current.getElement().clientWidth)
+      const offsetX = refTop.current.scrollLeft
+      const scrollLeftMax = (refTop.current.scrollWidth - refTop.current.clientWidth)
       dispatch(setCurrentlyAt((offsetX / scrollLeftMax) * (duration)))
     }
   }
@@ -96,15 +97,15 @@ import { t } from "i18next";
         css={{
           position: 'absolute',
           width: '2px',
-          height: '190px',
-          ...(refTop.current) && {left: (refTop.current.getElement().clientWidth / 2)},
+          height: '200px',
+          ...(refTop.current) && {left: (refTop.current.clientWidth / 2)},
           top: '10px',
           background: `${theme.text}`,
           zIndex: 100,
         }}
       />
       {/* Scrollable timeline container. Has width of parent*/}
-      <ScrollContainer ref={refTop} css={{overflow: 'hidden', width: '100%', height: '215px'}}
+      <ScrollContainer innerRef={refTop} css={{overflow: 'hidden', width: '100%', height: '215px'}}
         vertical={false}
         horizontal={true}
         onEndScroll={onEndScroll}
@@ -144,16 +145,16 @@ import { t } from "i18next";
 
 
 
-    // <div className="layoutRoot absoluteLayout">
-    //   {/* <Example /> */}
-    //   <Example2 />
-    //   {/* <TimelineSubtitleSegment
-    //   timelineWidth={width}
-    //   cue={{id: '42', text:"HI", startTime: 1000, endTime: 5000, tree:{children: [{type: "", value: ""}]}}}
-    //   index={0}
-    //   height={80}
-    //   /> */}
-    // </div>
+  // <div className="layoutRoot absoluteLayout">
+  //   {/* <Example /> */}
+  //   <Example2 />
+  //   {/* <TimelineSubtitleSegment
+  //   timelineWidth={width}
+  //   cue={{id: '42', text:"HI", startTime: 1000, endTime: 5000, tree:{children: [{type: "", value: ""}]}}}
+  //   index={0}
+  //   height={80}
+  //   /> */}
+  // </div>
   );
 };
 
@@ -164,7 +165,7 @@ import { t } from "i18next";
 const TimelineSubtitleSegmentsList: React.FC<{timelineWidth: number}> = ({timelineWidth}) => {
 
   const arbitraryHeight = 80
-  const subtitle = useSelector(selectSelectedSubtitleByFlavor)
+  const subtitle = useSelector(selectSelectedSubtitleById)
 
   const segmentsListStyle = css({
     position: 'relative',
@@ -175,7 +176,7 @@ const TimelineSubtitleSegmentsList: React.FC<{timelineWidth: number}> = ({timeli
 
   return (
     <div css={segmentsListStyle}>
-      {subtitle?.map((item, i) => {
+      {subtitle?.cues?.map((item, i) => {
         return (
           <TimelineSubtitleSegment timelineWidth={timelineWidth} cue={item} height={arbitraryHeight} key={item.idInternal} index={i}/>
         )
@@ -196,7 +197,7 @@ const TimelineSubtitleSegment: React.FC<{
 
   // Redux
   const dispatch = useDispatch()
-  const selectedFlavor = useSelector(selectSelectedSubtitleFlavor)
+  const selectedId = useSelector(selectSelectedSubtitleId)
   const duration = useSelector(selectDuration)
 
   // Dimensions and position offsets in px. Required for resizing
@@ -209,11 +210,11 @@ const TimelineSubtitleSegment: React.FC<{
   const [isGrabbed, setIsGrabbed] = useState(false)
   const nodeRef = React.useRef(null); // For supressing "ReactDOM.findDOMNode() is deprecated" warning
 
-  const theme = useSelector(selectTheme);
+  const theme = useTheme();
   // Reposition scrubber when the current x position was changed externally
   useEffect(() => {
     setControlledPosition({x: (props.cue.startTime / duration) * (props.timelineWidth), y: 0});
-  },[props.cue.startTime, duration, props.timelineWidth])
+  }, [props.cue.startTime, duration, props.timelineWidth])
 
   // Set width and reset any resizing that may have happened meanwhile
   useEffect(() => {
@@ -221,7 +222,7 @@ const TimelineSubtitleSegment: React.FC<{
     setAbsoluteHeight(props.height)
     setAbsoluteLeft(0)
     setAbsoluteTop(0)
-  },[duration, props.cue.endTime, props.cue.startTime, props.height, props.timelineWidth])
+  }, [duration, props.cue.endTime, props.cue.startTime, props.height, props.timelineWidth])
 
   // Check for impossible timestamps and update state in redux
   const dispatchNewTimes = (newStartTime: number, newEndTime: number) => {
@@ -233,7 +234,7 @@ const TimelineSubtitleSegment: React.FC<{
     }
 
     dispatch(setCueAtIndex({
-      identifier: selectedFlavor,
+      identifier: selectedId,
       cueIndex: props.index,
       newCue: {
         id: props.cue.id,
@@ -249,7 +250,7 @@ const TimelineSubtitleSegment: React.FC<{
   // Resizable does not support resizing in the west/north directions out of the box,
   // so additional calculations are necessary.
   // Adapted from Resizable example code
-  const onResizeAbsolute = (event: any, {element, size, handle}: any) => {
+  const onResizeAbsolute = (_event: any, {size, handle}: any) => {
     // Possible TODO: Find a way to stop resizing a segment beyond 0ms here instead of later
     let newLeft = absoluteLeft;
     let newTop = absoluteTop;
@@ -273,7 +274,7 @@ const TimelineSubtitleSegment: React.FC<{
   };
 
   // Update redux state based on the resize
-  const onResizeStop = (event: any, {element, size, handle}: any) => {
+  const onResizeStop = (_event: any, {handle}: any) => {
     // Calc new width, factoring in offset
     const newWidth = absoluteWidth
 
@@ -302,11 +303,11 @@ const TimelineSubtitleSegment: React.FC<{
     setAbsoluteTop(0)
   }
 
-  const onStartDrag = (e: DraggableEvent) => {
+  const onStartDrag = (_e: DraggableEvent) => {
     setIsGrabbed(true)
   }
 
-  const onStopDrag = (e: DraggableEvent, position: any) => {
+  const onStopDrag = (_e: DraggableEvent, position: any) => {
     // Update position and thereby start/end times in redux
     const {x} = position
     dispatchNewTimes(
@@ -357,32 +358,32 @@ const TimelineSubtitleSegment: React.FC<{
   })
 
   return (
-      <Draggable
-        onStart={onStartDrag}
-        onStop={onStopDrag}
-        defaultPosition={{ x: 10, y: 10 }}
-        position={controlledPosition}
-        axis="x"
-        bounds="parent"
-        nodeRef={nodeRef}
-        cancel={".react-resizable-handle"}
+    <Draggable
+      onStart={onStartDrag}
+      onStop={onStopDrag}
+      defaultPosition={{ x: 10, y: 10 }}
+      position={controlledPosition}
+      axis="x"
+      bounds="parent"
+      nodeRef={nodeRef}
+      cancel={".react-resizable-handle"}
+    >
+      <Resizable
+        height={absoluteHeight}
+        width={absoluteWidth}
+        onResize={onResizeAbsolute}
+        onResizeStop={onResizeStop}
+        // TODO: The 'e' handle is currently NOT WORKING CORRECTLY!
+        //  The errounous behaviour can already be seens with a minimal
+        //  draggable + resizable example.
+        //  Fix most likely requires changes in one of those modules
+        resizeHandles={['w']}
       >
-        <Resizable
-          height={absoluteHeight}
-          width={absoluteWidth}
-          onResize={onResizeAbsolute}
-          onResizeStop={onResizeStop}
-          // TODO: The 'e' handle is currently NOT WORKING CORRECTLY!
-          //  The errounous behaviour can already be seens with a minimal
-          //  draggable + resizable example.
-          //  Fix most likely requires changes in one of those modules
-          resizeHandles={['w']}
-        >
-          <div css={ segmentStyle } ref={nodeRef} onClick={onClick} id="no-scrolling">
-            <span css={textStyle}>{props.cue.text}</span>
-          </div>
-        </Resizable>
-      </Draggable>
+        <div css={segmentStyle} ref={nodeRef} onClick={onClick} id="no-scrolling">
+          <span css={textStyle}>{props.cue.text}</span>
+        </div>
+      </Resizable>
+    </Draggable>
   )
 })
 
@@ -390,7 +391,7 @@ const TimelineSubtitleSegment: React.FC<{
 //  * For debugging
 //  * Minimal example: Resizable
 //  */
-//  const Example: React.FC<{}> = () => {
+//  const Example: React.FC = () => {
 
 //   const [absoluteWidth, setAbsoluteWidth] = useState(200)
 //   const [absoluteHeight, setAbsoluteHeight] = useState(200)
@@ -447,7 +448,7 @@ const TimelineSubtitleSegment: React.FC<{
 //  * Minimal example: Draggable + Resizable
 //  * Erratic behaviour when resizing the east handle for smallish widths
 //  */
-// const Example2: React.FC<{}> = () => {
+// const Example2: React.FC = () => {
 
 //   const [absoluteWidth, setAbsoluteWidth] = useState(200)
 //   const [absoluteHeight, setAbsoluteHeight] = useState(200)
