@@ -2,36 +2,27 @@ import React from "react";
 
 import { css } from '@emotion/react'
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faCheckCircle, faTimesCircle } from "@fortawesome/free-solid-svg-icons";
+import { LuCheckCircle, LuXCircle} from "react-icons/lu";
 
 import { useSelector } from 'react-redux';
 import { selectEndState } from '../redux/endSlice'
 import { basicButtonStyle, flexGapReplacementStyle, navigationButtonStyle } from "../cssStyles";
 
-import './../i18n/config';
 import { useTranslation } from 'react-i18next';
-import { selectTheme } from "../redux/themeSlice";
+import { useTheme } from "../themes";
 import { ThemedTooltip } from "./Tooltip";
+import { CallbackButton } from "./Finish";
 
 /**
  * This page is to be displayed when the user is "done" with the editor
  * and should not be able to perfom any actions anymore
  */
-const TheEnd : React.FC<{}> = () => {
+const TheEnd : React.FC = () => {
 
   const { t } = useTranslation();
 
   // Init redux variables
   const endState = useSelector(selectEndState)
-
-  const icon = () => {
-    if (endState === 'discarded') {
-      return faTimesCircle
-    } else {
-      return faCheckCircle
-    }
-  }
 
   const text = () => {
     if (endState === 'discarded') {
@@ -43,7 +34,7 @@ const TheEnd : React.FC<{}> = () => {
 
   const theEndStyle = css({
     width: '100%',
-    height: '100%',
+    height: 'calc(100% - 64px)',
     display: 'flex',
     flexDirection: 'column',
     justifyContent: 'center',
@@ -51,20 +42,29 @@ const TheEnd : React.FC<{}> = () => {
     ...(flexGapReplacementStyle(20, false)),
   })
 
+  const restartOrBackStyle = css({
+    display: "flex",
+    flexDirection: 'row',
+    ...(flexGapReplacementStyle(20, false)),
+  })
+
   return (
     <div css={theEndStyle}>
-      <FontAwesomeIcon icon={icon()} size="10x" />
+      {endState === 'discarded' ? <LuXCircle css={{fontSize: 80}}/> : <LuCheckCircle css={{fontSize: 80}}/> }
       <div>{text()}</div>
-      {(endState === 'discarded') && <StartOverButton />}
+      <div css={restartOrBackStyle}>
+        <CallbackButton />
+        {(endState === 'discarded') && <StartOverButton />}
+      </div>
     </div>
   );
 }
 
 
-const StartOverButton: React.FC<{}> = () => {
+const StartOverButton: React.FC = () => {
 
   const { t } = useTranslation();
-  const theme = useSelector(selectTheme);
+  const theme = useTheme();
 
   const reloadPage = () => {
     window.location.reload();
@@ -74,11 +74,10 @@ const StartOverButton: React.FC<{}> = () => {
     <ThemedTooltip title={t("theEnd.startOver-tooltip")}>
       <div css={[basicButtonStyle(theme), navigationButtonStyle(theme)]}
         role="button" tabIndex={0}
-        onClick={ reloadPage }
+        onClick={reloadPage}
         onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
           reloadPage()
-        }}}>
-        {/* <FontAwesomeIcon icon={icon} spin={spin} size="1x"/> */}
+        } }}>
         <span>{t("theEnd.startOver-button")}</span>
       </div>
     </ThemedTooltip>
