@@ -7,11 +7,10 @@ import { css } from '@emotion/react'
 import { useSelector, useDispatch } from 'react-redux';
 import { Segment, httpRequestState } from '../types'
 import {
-  selectSegments, selectActiveSegmentIndex, selectDuration, selectVideoURL,  selectWaveformImages, setWaveformImages, selectTimelineZoom
+  selectSegments, selectActiveSegmentIndex, selectDuration, selectVideoURL, selectWaveformImages, setWaveformImages, selectTimelineZoom
 } from '../redux/videoSlice'
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBars, faSpinner } from "@fortawesome/free-solid-svg-icons";
+import { LuMenu, LuLoader} from "react-icons/lu";
 
 import useResizeObserver from "use-resize-observer";
 
@@ -20,13 +19,13 @@ import { convertMsToReadableString } from '../util/utilityFunctions';
 import { GlobalHotKeys } from 'react-hotkeys';
 import { scrubberKeyMap } from '../globalKeys';
 
-import './../i18n/config';
 import { useTranslation } from 'react-i18next';
 import { ActionCreatorWithPayload } from '@reduxjs/toolkit';
 import { RootState } from '../redux/store';
-import { selectTheme } from '../redux/themeSlice';
+import { useTheme } from "../themes";
 import { ThemedTooltip } from './Tooltip';
 import ScrollContainer from 'react-indiana-drag-scroll';
+import { spinningStyle } from '../cssStyles';
 
 /**
  * A container for visualizing the cutting of the video, as well as for controlling
@@ -43,7 +42,7 @@ const Timeline: React.FC<{
   setCurrentlyAt: ActionCreatorWithPayload<number, string>,
   setIsPlaying: ActionCreatorWithPayload<boolean, string>,
 }> = ({
-  timelineHeight = 250,
+  timelineHeight = 200,
   styleByActiveSegment = true,
   selectCurrentlyAt,
   selectIsPlaying,
@@ -57,8 +56,8 @@ const Timeline: React.FC<{
   const duration = useSelector(selectDuration)
   const timelineZoom = useSelector(selectTimelineZoom)
 
-  const { ref, width = 1, } = useResizeObserver<HTMLDivElement>();
-  const scrollContainerRef = useRef<ScrollContainer>(null);
+  const { ref, width = 1 } = useResizeObserver<HTMLDivElement>();
+  const scrollContainerRef = useRef(null);
 
   const timelineStyle = css({
     position: 'relative',     // Need to set position for Draggable bounds to work
@@ -68,36 +67,36 @@ const Timeline: React.FC<{
 
   // Update the current time based on the position clicked on the timeline
   const setCurrentlyAtToClick = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    let rect = e.currentTarget.getBoundingClientRect()
-    let offsetX = e.clientX - rect.left
+    const rect = e.currentTarget.getBoundingClientRect()
+    const offsetX = e.clientX - rect.left
     dispatch(setClickTriggered(true))
     dispatch(setCurrentlyAt((offsetX / width) * (duration)))
   }
 
   return (
-  <ScrollContainer ref={scrollContainerRef} css={{overflowY: 'hidden', width: '100%', height: `${timelineHeight}px`}}
-    vertical={false}
-    horizontal={true}
-    ignoreElements={"#no-scrolling"}  // dom elements with this id in the container will not trigger scrolling when dragged
-    hideScrollbars={false}            // ScrollContainer hides scrollbars per default
-  >
-    <div ref={ref} css={timelineStyle} onMouseUp={e => setCurrentlyAtToClick(e)}>
-      <Scrubber
-        timelineWidth={width}
-        timelineHeight={timelineHeight}
-        selectCurrentlyAt={selectCurrentlyAt}
-        selectIsPlaying={selectIsPlaying}
-        setCurrentlyAt={setCurrentlyAt}
-        setIsPlaying={setIsPlaying}
-      />
+    <ScrollContainer innerRef={scrollContainerRef} css={{overflowY: 'hidden', width: '100%', height: `${timelineHeight}px`}}
+      vertical={false}
+      horizontal={true}
+      ignoreElements={"#no-scrolling"}  // dom elements with this id in the container will not trigger scrolling when dragged
+      hideScrollbars={false}            // ScrollContainer hides scrollbars per default
+    >
+      <div ref={ref} css={timelineStyle} onMouseUp={e => setCurrentlyAtToClick(e)}>
+        <Scrubber
+          timelineWidth={width}
+          timelineHeight={timelineHeight}
+          selectCurrentlyAt={selectCurrentlyAt}
+          selectIsPlaying={selectIsPlaying}
+          setCurrentlyAt={setCurrentlyAt}
+          setIsPlaying={setIsPlaying}
+        />
 
-        <div css={{position: 'relative', height: timelineHeight - 20 + 'px'}} >
+        <div css={{position: 'relative', height: timelineHeight + 'px'}} >
           <Waveforms timelineHeight={timelineHeight}/>
           <SegmentsList timelineWidth={width} timelineHeight={timelineHeight} styleByActiveSegment={styleByActiveSegment} tabable={true}/>
         </div>
 
-    </div>
-  </ScrollContainer>
+      </div>
+    </ScrollContainer>
   );
 };
 
@@ -130,10 +129,10 @@ export const Scrubber: React.FC<{
   const duration = useSelector(selectDuration)
   const activeSegmentIndex = useSelector(selectActiveSegmentIndex)  // For ARIA information display
   const segments = useSelector(selectSegments)                      // For ARIA information display
-  const theme = useSelector(selectTheme)
+  const theme = useTheme()
 
   // Init state variables
-  const [controlledPosition, setControlledPosition] = useState({x: 0,y: 0,});
+  const [controlledPosition, setControlledPosition] = useState({ x: 0, y: 0 });
   const [isGrabbed, setIsGrabbed] = useState(false)
   const [wasPlayingWhenGrabbed, setWasPlayingWhenGrabbed] = useState(false)
   const [keyboardJumpDelta, setKeyboardJumpDelta] = useState(1000)  // In milliseconds. For keyboard navigation
@@ -142,7 +141,7 @@ export const Scrubber: React.FC<{
 
   // Reposition scrubber when the current x position was changed externally
   useEffect(() => {
-    if(currentlyAt !== wasCurrentlyAtRef.current && !isGrabbed) {
+    if (currentlyAt !== wasCurrentlyAtRef.current && !isGrabbed) {
       updateXPos();
       wasCurrentlyAtRef.current = currentlyAt;
     }
@@ -150,14 +149,14 @@ export const Scrubber: React.FC<{
 
   // Reposition scrubber when the timeline width changes
   useEffect(() => {
-    if(currentlyAt && duration) {
+    if (currentlyAt && duration) {
       setControlledPosition({x: (currentlyAt / duration) * (timelineWidth), y: 0});
     }
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [timelineWidth])
 
   // Callback for when the scrubber gets dragged by the user
-  const onControlledDrag = (e: any, position: any) => {
+  const onControlledDrag = (_e: any, position: any) => {
     // Update position
     const {x} = position
     dispatch(setCurrentlyAt((x / timelineWidth) * (duration)))
@@ -180,7 +179,7 @@ export const Scrubber: React.FC<{
     }
   }
 
-  const onStopDrag = (e: any, position: any) => {
+  const onStopDrag = (_e: any, position: any) => {
     // Update position
     const {x} = position;
     setControlledPosition({x, y: 0});
@@ -204,56 +203,46 @@ export const Scrubber: React.FC<{
   }
 
   const scrubberStyle = css({
-    backgroundColor: `${theme.text}`,
-    height: timelineHeight - 10 + 'px', //    TODO: CHECK IF height: '100%',
+    backgroundColor: `${theme.scrubber}`,
+    height: timelineHeight + 20 + 'px', //    TODO: CHECK IF height: '100%',
     width: '1px',
     position: 'absolute',
     zIndex: 2,
-    boxShadow: `${theme.boxShadow}`,
     display: 'flex',
     flexDirection: 'column',
     justifyContent: 'space-between',
     alignItems: 'center',
-    outline: `${theme.scrubber}`,
+    top: '-20px',
   });
 
   const scrubberDragHandleStyle = css({
     // Base style
-    background: `${theme.text}`,
+    background: `${theme.scrubber_handle}`,
     display: "inline-block",
-    height: "10px",
+    height: "20px",
     position: "relative",
     width: "20px",
+    borderRadius: '5px',
+    boxShadow: `${theme.boxShadow_tiles}`,
     "&:after": {
-      borderTop: `10px solid ${theme.text}`,
+      borderTop: `10px solid ${theme.scrubber_handle}`,
       borderLeft: '10px solid transparent',
       borderRight: '10px solid transparent',
       content: '""',
       height: 0,
       left: 0,
       position: "absolute",
-      top: "10px",
+      top: "17px",
       width: 0,
     },
     // Animation
     cursor: isGrabbed ? "grabbing" : "grab",
-    transitionDuration: "0.3s",
-    transitionProperty: "transform",
-    "&:hover": {
-      transform: 'scale(1.1)',
-    },
-    "&:focus": {
-      transform: 'scale(1.1)',
-    },
-    "&:active": {
-      transform: 'scale(0.9)',
-    },
   })
 
   const scrubberDragHandleIconStyle = css({
-    transform: 'scaleY(0.7) rotate(90deg)',
-    paddingRight: '5px',
-    color: `${theme.background}`,
+    paddingLeft: '2px',
+    paddingTop: '2px',
+    color: `${theme.scrubber_icon}`,
   })
 
   // // Possible TODO: Find a way to use ariaLive in a way that only the latest change is announced
@@ -271,21 +260,20 @@ export const Scrubber: React.FC<{
         bounds="parent"
         position={controlledPosition}
         nodeRef={nodeRef}
-        >
-          <div ref={nodeRef} css={scrubberStyle} id="no-scrolling">
-
-            <div css={scrubberDragHandleStyle} aria-grabbed={isGrabbed}
-              aria-label={t("timeline.scrubber-text-aria",
-                         {currentTime: convertMsToReadableString(currentlyAt), segment: activeSegmentIndex,
-                          segmentStatus: (segments[activeSegmentIndex].deleted ? "Deleted" : "Alive"),
-                          moveLeft: scrubberKeyMap[handlers.left.name],
-                          moveRight: scrubberKeyMap[handlers.right.name],
-                          increase: scrubberKeyMap[handlers.increase.name],
-                          decrease: scrubberKeyMap[handlers.decrease.name] })}
-              tabIndex={0}>
-              <FontAwesomeIcon css={scrubberDragHandleIconStyle} icon={faBars} size="1x" />
-            </div>
+      >
+        <div ref={nodeRef} css={scrubberStyle} id="no-scrolling">
+          <div css={scrubberDragHandleStyle} aria-grabbed={isGrabbed}
+            aria-label={t("timeline.scrubber-text-aria",
+              {currentTime: convertMsToReadableString(currentlyAt), segment: activeSegmentIndex,
+                segmentStatus: (segments[activeSegmentIndex].deleted ? "Deleted" : "Alive"),
+                moveLeft: scrubberKeyMap[handlers.left.name],
+                moveRight: scrubberKeyMap[handlers.right.name],
+                increase: scrubberKeyMap[handlers.increase.name],
+                decrease: scrubberKeyMap[handlers.decrease.name] })}
+            tabIndex={0}>
+            <LuMenu css={scrubberDragHandleIconStyle}/>
           </div>
+        </div>
       </Draggable>
     </GlobalHotKeys>
   );
@@ -300,7 +288,6 @@ export const SegmentsList: React.FC<{
   styleByActiveSegment?: boolean,
   tabable?: boolean,
 }> = ({
-  timelineWidth,
   timelineHeight,
   styleByActiveSegment = true,
   tabable = true,
@@ -319,49 +306,48 @@ export const SegmentsList: React.FC<{
    */
   const bgColor = (deleted: boolean, active: boolean) => {
     if (!deleted && !active) {
-      return 'rgba(0, 0, 255, 0.4)'
+      return 'rgba(137, 137, 137, 0.4)'
     } else if (deleted && !active) {
       return `repeating-linear-gradient(
-                -45deg,
-                rgba(255, 45, 45, 0.4),
-                rgba(255, 45, 45, 0.4) 10px,
-                rgba(255, 0, 0, 0.4) 10px,
-                rgba(255, 0, 0, 0.4) 20px);`
+                -35deg,
+                rgba(200, 0, 0, 0.4),
+                rgba(200, 0, 0, 0.4) 2px,
+                rgba(255, 95, 95, 0.4) 2px,
+                rgba(255, 95, 95, 0.4) 50px);`
     } else if (!deleted && active) {
-      return 'rgba(0, 0, 200, 0.4)'
+      return 'rgba(78, 78, 78, 0.4)'
     } else if (deleted && active) {
       return `repeating-linear-gradient(
-                -45deg,
-                rgba(200, 45, 45, 0.4),
-                rgba(200, 45, 45, 0.4) 10px,
-                rgba(200, 0, 0, 0.4) 10px,
-                rgba(200, 0, 0, 0.4) 20px);`
+                -35deg,
+                rgba(180, 0, 0, 0.4),
+                rgba(180, 0, 0, 0.4) 2px,
+                rgba(255, 65, 65, 0.4) 2px,
+                rgba(255, 65, 65, 0.4) 50px);`
     }
   }
 
   // Render the individual segments
   const renderedSegments = () => {
     return (
-      segments.map( (segment: Segment, index: number) => (
-        <ThemedTooltip title={t("timeline.segment-tooltip", {segment: index})}>
-          <div key={segment.id}
+      segments.map((segment: Segment, index: number) => (
+        <ThemedTooltip title={t("timeline.segment-tooltip", {segment: index})} key={segment.id}>
+          <div
             aria-label={t("timeline.segments-text-aria",
-                      {segment: index,
-                        segmentStatus: (segment.deleted ? "Deleted" : "Alive"),
-                        start: convertMsToReadableString(segment.start),
-                        end: convertMsToReadableString(segment.end) })}
+              {segment: index,
+                segmentStatus: (segment.deleted ? "Deleted" : "Alive"),
+                start: convertMsToReadableString(segment.start),
+                end: convertMsToReadableString(segment.end) })}
             tabIndex={tabable ? 0 : -1}
-          css={{
-            background: bgColor(segment.deleted, styleByActiveSegment ? activeSegmentIndex === index : false),
-            borderRadius: '5px',
-            borderStyle: styleByActiveSegment ? (activeSegmentIndex === index ? 'dashed' : 'solid') : 'solid',
-            borderColor: 'white',
-            borderWidth: '1px',
-            boxSizing: 'border-box',
-            width: ((segment.end - segment.start) / duration) * 100 + '%',
-            height: timelineHeight - 20 + 'px',     // CHECK IF 100%
-            zIndex: 1,
-          }}>
+            css={{
+              background: bgColor(segment.deleted, styleByActiveSegment ? activeSegmentIndex === index : false),
+              borderStyle: styleByActiveSegment ? (activeSegmentIndex === index ? 'dashed' : 'solid') : 'solid',
+              borderColor: 'white',
+              borderWidth: '1px',
+              boxSizing: 'border-box',
+              width: ((segment.end - segment.start) / duration) * 100 + '%',
+              height: timelineHeight + 'px',     // CHECK IF 100%
+              zIndex: 1,
+            }}>
           </div>
         </ThemedTooltip>
       ))
@@ -371,7 +357,7 @@ export const SegmentsList: React.FC<{
   const segmentsStyle = css({
     display: 'flex',
     flexDirection: 'row',
-    paddingTop: '10px',
+    // paddingTop: '10px',
     height: '100%',
   })
 
@@ -392,7 +378,7 @@ export const Waveforms: React.FC<{timelineHeight: number}> = ({timelineHeight}) 
   const dispatch = useDispatch();
   const videoURLs = useSelector(selectVideoURL)
   const videoURLStatus = useSelector((state: { videoState: { status: httpRequestState["status"] } }) => state.videoState.status);
-  const theme = useSelector(selectTheme);
+  const theme = useTheme();
 
   // Update based on current fetching status
   const images = useSelector(selectWaveformImages)
@@ -401,56 +387,54 @@ export const Waveforms: React.FC<{timelineHeight: number}> = ({timelineHeight}) 
   const waveformDisplayTestStyle = css({
     display: 'flex',
     flexDirection: 'column',
-    position: "absolute" as "absolute",
+    position: "absolute" as const,
     justifyContent: 'center',
     ...(images.length <= 0) && {alignItems: 'center'},  // Only center during loading
     width: '100%',
-    height: timelineHeight - 20 + 'px',   // CHECK IF     height: '100%',
-    paddingTop: '10px',
+    height: timelineHeight + 'px',   // CHECK IF     height: '100%',
+    // paddingTop: '10px',
     filter: `${theme.invert_wave}`,
     color: `${theme.inverted_text}`,
   });
 
   const waveformStyle = css({
     background: `${theme.waveform_bg}`,
-    filter: `${theme.waveform_filter}`,
     borderRadius: '5px',
   });
 
   // When the URLs to the videos are fetched, generate waveforms
-  useEffect( () => {
+  useEffect(() => {
     if (videoURLStatus === 'success') {
       if (images.length > 0) {
         return
       }
 
       const newImages: string[] = []    // Store local paths to image files
-      let waveformsProcessed : number = 0  // Counter for checking if all workers are done
+      let waveformsProcessed = 0  // Counter for checking if all workers are done
 
       // Only display the waveform of the first video we get
       const onlyOneURL = [videoURLs[0]]
 
       onlyOneURL.forEach((videoURL, _index, array) => {
         // Set up blob request
-        var blob = null
-        var xhr = new XMLHttpRequest()
+        let blob = null
+        const xhr = new XMLHttpRequest()
         xhr.open("GET", videoURL)
         xhr.responseType = "blob"
-        xhr.onload = function()
-        {
+        xhr.onload = () => {
           blob = xhr.response
-          var file = new File([blob], blob)
+          const file = new File([blob], blob)
 
           // Start waveform worker with blob
           const waveformWorker : any = new Waveform({type: 'img', width: '2000', height: '230', samples: 100000, media: file})
 
-          waveformWorker.onerror = function(error: string) {
+          waveformWorker.onerror = (error: string) => {
             setWaveformWorkerError(true)
             console.log("Waveform could not be generated:" + error)
           }
 
           // When done, save path to generated waveform img
-          waveformWorker.oncomplete = function(image: any, numSamples: any) {
+          waveformWorker.oncomplete = (image: any, _numSamples: any) => {
             newImages.push(image)
             waveformsProcessed++
             // If all images are generated, rerender
@@ -469,7 +453,7 @@ export const Waveforms: React.FC<{timelineHeight: number}> = ({timelineHeight}) 
   const renderImages = () => {
     if (images.length > 0) {
       return (
-        <img alt='Waveform' src={images[0]} css={[waveformStyle, {minHeight: 0, height: '100%'}]}></img>
+        <img alt="Waveform" src={images[0]} css={[waveformStyle, {minHeight: 0, height: '100%'}]}></img>
         // images.map((image, index) =>
         //   <img key={index} alt='Waveform' src={image ? image : ""} css={{minHeight: 0}}></img>
         // )
@@ -483,7 +467,7 @@ export const Waveforms: React.FC<{timelineHeight: number}> = ({timelineHeight}) 
     else {
       return (
         <>
-          <FontAwesomeIcon icon={faSpinner} spin size="3x"/>
+          <LuLoader css={[spinningStyle, {fontSize: 40}]}/>
           <div>{t("timeline.generateWaveform-text")}</div>
         </>
       );
@@ -491,9 +475,9 @@ export const Waveforms: React.FC<{timelineHeight: number}> = ({timelineHeight}) 
   }
 
   return (
-  <div css={waveformDisplayTestStyle}>
-    {renderImages()}
-  </div>
+    <div css={waveformDisplayTestStyle}>
+      {renderImages()}
+    </div>
   );
 }
 
