@@ -1,11 +1,9 @@
 import React, { useEffect, useState } from "react";
 
 import { css } from '@emotion/react'
-import { basicButtonStyle, backOrContinueStyle, errorBoxStyle, flexGapReplacementStyle } from '../cssStyles'
+import { basicButtonStyle, backOrContinueStyle, errorBoxStyle, flexGapReplacementStyle, spinningStyle } from '../cssStyles'
 
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faTools} from "@fortawesome/free-solid-svg-icons";
-import { faSpinner, faCheck, faExclamationCircle, faChevronLeft, faFileExport } from "@fortawesome/free-solid-svg-icons";
+import { LuLoader, LuCheck, LuAlertCircle, LuChevronLeft, LuDatabase, LuMoreHorizontal} from "react-icons/lu";
 
 import { useDispatch, useSelector } from 'react-redux';
 import { selectSegments, selectTracks, setHasChanges as videoSetHasChanges, selectSelectedWorkflowId } from '../redux/videoSlice'
@@ -14,19 +12,17 @@ import { postVideoInformationWithWorkflow, selectStatus, selectError } from '../
 import { PageButton } from './Finish'
 import { setEnd } from "../redux/endSlice";
 
-import './../i18n/config';
 import { useTranslation } from 'react-i18next';
 import { postMetadata, selectPostError, selectPostStatus, setHasChanges as metadataSetHasChanges } from "../redux/metadataSlice";
 import { AppDispatch } from "../redux/store";
 import { selectSubtitles } from "../redux/subtitleSlice";
 import { serializeSubtitle } from "../util/utilityFunctions";
-import { Flavor } from "../types";
-import { selectTheme } from "../redux/themeSlice";
+import { useTheme } from "../themes";
 
 /**
  * Will eventually display settings based on the selected workflow index
  */
-const WorkflowConfiguration : React.FC<{}> = () => {
+const WorkflowConfiguration : React.FC = () => {
 
   const { t } = useTranslation();
 
@@ -34,7 +30,7 @@ const WorkflowConfiguration : React.FC<{}> = () => {
   const postAndProcessError = useSelector(selectError)
   const postMetadataStatus = useSelector(selectPostStatus);
   const postMetadataError = useSelector(selectPostError);
-  const theme = useSelector(selectTheme);
+  const theme = useTheme();
 
   const workflowConfigurationStyle = css({
     display: 'flex',
@@ -47,20 +43,20 @@ const WorkflowConfiguration : React.FC<{}> = () => {
   return (
     <div css={workflowConfigurationStyle}>
       <h2>{t("workflowConfig.headline-text")}</h2>
-      <FontAwesomeIcon icon={faTools} size="10x" />
+      <LuMoreHorizontal css={{fontSize: 80}} />
       Placeholder
       <div>{t("workflowConfig.satisfied-text")}</div>
       <div css={backOrContinueStyle}>
-        <PageButton pageNumber={1} label={t("various.goBack-button")} iconName={faChevronLeft}/>
+        <PageButton pageNumber={1} label={t("various.goBack-button")} Icon={LuChevronLeft}/>
         <SaveAndProcessButton text={t("workflowConfig.confirm-button")}/>
       </div>
       <div css={errorBoxStyle(postAndProcessWorkflowStatus === "failed", theme)} role="alert">
         <span>{t("various.error-text")}</span><br />
-        {postAndProcessError ? t("various.error-details-text", {errorMessage: postAndProcessError}) : t("various.error-noDetails-text")}<br/>
+        {postAndProcessError ? t("various.error-details-text", {errorMessage: postAndProcessError}) : t("various.error-text")}<br/>
       </div>
       <div css={errorBoxStyle(postMetadataStatus === "failed", theme)} role="alert">
         <span>{t("various.error-text")}</span><br />
-        {postMetadataError ? t("various.error-details-text", {errorMessage: postMetadataError}) : t("various.error-noDetails-text")}<br />
+        {postMetadataError ? t("various.error-details-text", {errorMessage: postMetadataError}) : t("various.error-text")}<br />
       </div>
     </div>
   );
@@ -82,7 +78,7 @@ export const SaveAndProcessButton: React.FC<{text: string}> = ({text}) => {
   const workflowStatus = useSelector(selectStatus);
   const metadataStatus = useSelector(selectPostStatus);
   const [metadataSaveStarted, setMetadataSaveStarted] = useState(false);
-  const theme = useSelector(selectTheme);
+  const theme = useTheme();
 
   // Let users leave the page without warning after a successful save
   useEffect(() => {
@@ -97,9 +93,11 @@ export const SaveAndProcessButton: React.FC<{text: string}> = ({text}) => {
     const subtitlesForPosting = []
 
     for (const identifier in subtitles) {
-      let flavor: Flavor = {type: identifier.split("/")[0], subtype: identifier.split("/")[1]}
-      subtitlesForPosting.push({flavor: flavor, subtitle: serializeSubtitle(subtitles[identifier])})
-
+      subtitlesForPosting.push({
+        id: identifier,
+        subtitle: serializeSubtitle(subtitles[identifier].cues),
+        tags: subtitles[identifier].tags
+      })
     }
     return subtitlesForPosting
   }
@@ -127,16 +125,16 @@ export const SaveAndProcessButton: React.FC<{text: string}> = ({text}) => {
   }, [metadataStatus])
 
   // Update based on current fetching status
-  let icon = faFileExport
+  let Icon = LuDatabase
   let spin = false
   if (workflowStatus === 'failed' || metadataStatus === 'failed') {
-    icon = faExclamationCircle
+    Icon = LuAlertCircle
     spin = false
   } else if (workflowStatus === 'success' && metadataStatus === 'success') {
-    icon = faCheck
+    Icon = LuCheck
     spin = false
   } else if (workflowStatus === 'loading' || metadataStatus === 'loading') {
-    icon = faSpinner
+    Icon = LuLoader
     spin = true
 
   }
@@ -150,11 +148,11 @@ export const SaveAndProcessButton: React.FC<{text: string}> = ({text}) => {
   return (
     <div css={[basicButtonStyle(theme), saveButtonStyle]}
       role="button" tabIndex={0}
-      onClick={ saveAndProcess }
+      onClick={saveAndProcess}
       onKeyDown={(event: React.KeyboardEvent<HTMLDivElement>) => { if (event.key === " " || event.key === "Enter") {
         saveAndProcess()
-      }}}>
-      <FontAwesomeIcon  icon={icon} spin={spin} size="1x"/>
+      } }}>
+      <Icon css={spin ? spinningStyle : undefined}/>
       <span>{text}</span>
     </div>
   );
