@@ -1,12 +1,12 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import { IconDefinition } from "@fortawesome/free-solid-svg-icons";
+import { IconType } from "react-icons";
 
 interface error {
   error: boolean,
-  errorTitle: string | undefined,
+  errorTitle?: string,
   errorMessage: string,
-  errorDetails: string | undefined,
-  errorIcon: IconDefinition | undefined,
+  errorDetails?: string,
+  errorIcon?: IconType,
 }
 
 const initialState: error = {
@@ -26,10 +26,10 @@ export const errorSlice = createSlice({
   reducers: {
     setError: (state, action: PayloadAction<{
       error: error["error"],
-      errorTitle: error["errorTitle"],
+      errorTitle?: error["errorTitle"],
       errorMessage: error["errorMessage"],
-      errorDetails: error["errorDetails"],
-      errorIcon: error["errorIcon"]
+      errorDetails?: error["errorDetails"],
+      errorIcon?: error["errorIcon"]
     }>) => {
       state.error = action.payload.error;
       state.errorTitle = action.payload.errorTitle;
@@ -40,7 +40,7 @@ export const errorSlice = createSlice({
   }
 })
 
-export const { setError, } = errorSlice.actions
+export const { setError } = errorSlice.actions
 
 // Export Selectors
 export const selectIsError = (state: { errorState: { error: error["error"] }; }) => state.errorState.error
