@@ -23,7 +23,7 @@ export const mainMenuSlice = createSlice({
   }
 })
 
-export const { setState, } = mainMenuSlice.actions
+export const { setState } = mainMenuSlice.actions
 
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
