@@ -1,8 +1,7 @@
-import { Segment, SubtitleCue } from './../types';
+import { Segment, SubtitleCue, SubtitlesInEditor } from './../types';
 import { createSlice, Dispatch, nanoid, PayloadAction } from '@reduxjs/toolkit'
 import { roundToDecimalPlace } from '../util/utilityFunctions';
 import type { RootState } from '../redux/store'
-import { WritableDraft } from 'immer/dist/internal';
 import { video } from './videoSlice';
 
 export interface subtitle {
@@ -12,12 +11,14 @@ export interface subtitle {
   previewTriggered: boolean,      // Basically acts as a callback for the video players.
   currentlyAt: number,            // Position in the video in milliseconds
   clickTriggered: boolean,        // Another video player callback
-  subtitles: { [identifier: string]: SubtitleCue[] },
-  selectedSubtitleFlavor: string,
+  subtitles: { [identifier: string]: SubtitlesInEditor },
+  selectedSubtitleId: string,
   aspectRatios: {width: number, height: number}[],  // Aspect ratios of every video
   focusSegmentTriggered: boolean,   // a segment in the timeline was clicked
   focusSegmentId: string,           // which segment in the timeline was clicked
   focusSegmentTriggered2: boolean,   // a different trigger for a child component, to avoid additional rerenders from the parent
+
+  hasChanges: boolean         // Did user make changes to metadata view since last save
 }
 
 const initialState: subtitle = {
@@ -28,12 +29,13 @@ const initialState: subtitle = {
   currentlyAt: 0,
   clickTriggered: false,
   subtitles: {},
-  selectedSubtitleFlavor: "",
+  selectedSubtitleId: "",
   focusSegmentTriggered: false,
   focusSegmentId: "",
   focusSegmentTriggered2: false,
 
   aspectRatios: [],
+  hasChanges: false,
 }
 
 const updateCurrentlyAt = (state: subtitle, milliseconds: number) => {
@@ -72,16 +74,16 @@ export const subtitleSlice = createSlice({
     setClickTriggered: (state, action) => {
       state.clickTriggered = action.payload
     },
-    setSubtitle: (state, action: PayloadAction<{identifier: string, subtitles: SubtitleCue[]}>) => {
+    setSubtitle: (state, action: PayloadAction<{identifier: string, subtitles: SubtitlesInEditor}>) => {
       state.subtitles[action.payload.identifier] = action.payload.subtitles
     },
     setCueAtIndex: (state, action: PayloadAction<{identifier: string, cueIndex: number, newCue: SubtitleCue}>) => {
-      if (action.payload.cueIndex < 0 || action.payload.cueIndex >= state.subtitles[action.payload.identifier].length) {
+      if (action.payload.cueIndex < 0 || action.payload.cueIndex >= state.subtitles[action.payload.identifier].cues.length) {
         console.log("WARNING: Tried to set segment for subtitle " + action.payload.identifier + " but was out of range")
         return
       }
 
-      let cue = state.subtitles[action.payload.identifier][action.payload.cueIndex]
+      const cue = state.subtitles[action.payload.identifier].cues[action.payload.cueIndex]
       cue.id = action.payload.newCue.id
       cue.idInternal = action.payload.newCue.idInternal
       cue.text = action.payload.newCue.text
@@ -90,9 +92,10 @@ export const subtitleSlice = createSlice({
 
       cue.tree.children[0].value = action.payload.newCue.text
 
-      state.subtitles[action.payload.identifier][action.payload.cueIndex] = cue
+      state.subtitles[action.payload.identifier].cues[action.payload.cueIndex] = cue
 
       sortSubtitle(state, action.payload.identifier)
+      state.hasChanges = true
     },
     addCueAtIndex: (state, action: PayloadAction<{identifier: string, cueIndex: number, text: string, startTime: number, endTime: number}>) => {
       const startTime = action.payload.startTime >= 0 ? action.payload.startTime : 0
@@ -110,30 +113,32 @@ export const subtitleSlice = createSlice({
       state.focusSegmentTriggered2 = true
       state.focusSegmentId = cue.idInternal
 
-      if (action.payload.cueIndex < 0 ) {
-        state.subtitles[action.payload.identifier].splice(0, 0, cue);
+      if (action.payload.cueIndex < 0) {
+        state.subtitles[action.payload.identifier].cues.splice(0, 0, cue);
       }
 
-      if (action.payload.cueIndex >= 0 || action.payload.cueIndex < state.subtitles[action.payload.identifier].length) {
-        state.subtitles[action.payload.identifier].splice(action.payload.cueIndex, 0, cue);
+      if (action.payload.cueIndex >= 0 || action.payload.cueIndex < state.subtitles[action.payload.identifier].cues.length) {
+        state.subtitles[action.payload.identifier].cues.splice(action.payload.cueIndex, 0, cue);
       }
 
-      if (action.payload.cueIndex >= state.subtitles[action.payload.identifier].length) {
-        state.subtitles[action.payload.identifier].push(cue)
+      if (action.payload.cueIndex >= state.subtitles[action.payload.identifier].cues.length) {
+        state.subtitles[action.payload.identifier].cues.push(cue)
       }
 
       sortSubtitle(state, action.payload.identifier)
+      state.hasChanges = true
     },
     removeCue: (state, action: PayloadAction<{identifier: string, cue: SubtitleCue}>) => {
-      const cueIndex = state.subtitles[action.payload.identifier].findIndex(i => i.idInternal === action.payload.cue.idInternal);
+      const cueIndex = state.subtitles[action.payload.identifier].cues.findIndex(i => i.idInternal === action.payload.cue.idInternal);
       if (cueIndex > -1) {
-        state.subtitles[action.payload.identifier].splice(cueIndex, 1);
+        state.subtitles[action.payload.identifier].cues.splice(cueIndex, 1);
       }
 
       sortSubtitle(state, action.payload.identifier)
+      state.hasChanges = true
     },
-    setSelectedSubtitleFlavor: (state, action: PayloadAction<subtitle["selectedSubtitleFlavor"]>) => {
-      state.selectedSubtitleFlavor = action.payload
+    setSelectedSubtitleId: (state, action: PayloadAction<subtitle["selectedSubtitleId"]>) => {
+      state.selectedSubtitleId = action.payload
     },
     setFocusSegmentTriggered: (state, action: PayloadAction<subtitle["focusSegmentTriggered"]>) => {
       state.focusSegmentTriggered = action.payload
@@ -146,37 +151,40 @@ export const subtitleSlice = createSlice({
       state.focusSegmentTriggered2 = action.payload
     },
     setFocusToSegmentAboveId: (state, action: PayloadAction<{identifier: string, segmentId: subtitle["focusSegmentId"]}>) => {
-      let cueIndex = state.subtitles[action.payload.identifier].findIndex(i => i.idInternal === action.payload.segmentId);
+      let cueIndex = state.subtitles[action.payload.identifier].cues.findIndex(i => i.idInternal === action.payload.segmentId);
       cueIndex = cueIndex - 1
-      if (cueIndex < 0 ) {
+      if (cueIndex < 0) {
         cueIndex = 0
       }
-      state.focusSegmentId = state.subtitles[action.payload.identifier][cueIndex].idInternal
+      state.focusSegmentId = state.subtitles[action.payload.identifier].cues[cueIndex].idInternal
     },
     setFocusToSegmentBelowId: (state, action: PayloadAction<{identifier: string, segmentId: subtitle["focusSegmentId"]}>) => {
-      let cueIndex = state.subtitles[action.payload.identifier].findIndex(i => i.idInternal === action.payload.segmentId);
+      let cueIndex = state.subtitles[action.payload.identifier].cues.findIndex(i => i.idInternal === action.payload.segmentId);
       cueIndex = cueIndex + 1
-      if (cueIndex >= state.subtitles[action.payload.identifier].length) {
-        cueIndex = state.subtitles[action.payload.identifier].length - 1
+      if (cueIndex >= state.subtitles[action.payload.identifier].cues.length) {
+        cueIndex = state.subtitles[action.payload.identifier].cues.length - 1
       }
-      state.focusSegmentId = state.subtitles[action.payload.identifier][cueIndex].idInternal
+      state.focusSegmentId = state.subtitles[action.payload.identifier].cues[cueIndex].idInternal
     },
-    setAspectRatio: (state, action: PayloadAction<{dataKey: number} & {width: number, height: number}> ) => {
+    setAspectRatio: (state, action: PayloadAction<{dataKey: number} & {width: number, height: number}>) => {
       state.aspectRatios[action.payload.dataKey] = {width: action.payload.width, height: action.payload.height}
+    },
+    setHasChanges: (state, action: PayloadAction<subtitle["hasChanges"]>) => {
+      state.hasChanges = action.payload
     },
   },
 })
 
 // Sort a subtitle array by startTime
-const sortSubtitle = (state: WritableDraft<subtitle>, identifier: string) => {
-  state.subtitles[identifier].sort((a, b) => a.startTime - b.startTime)
+const sortSubtitle = (state: subtitle, identifier: string) => {
+  state.subtitles[identifier].cues.sort((a, b) => a.startTime - b.startTime)
 }
 
 // Export Actions
 export const { setIsDisplayEditView, setIsPlaying, setIsPlayPreview, setPreviewTriggered, setCurrentlyAt,
   setCurrentlyAtInSeconds, setClickTriggered, setSubtitle, setCueAtIndex, addCueAtIndex, removeCue,
-  setSelectedSubtitleFlavor, setFocusSegmentTriggered, setFocusSegmentId, setFocusSegmentTriggered2,
-  setFocusToSegmentAboveId, setFocusToSegmentBelowId, setAspectRatio } = subtitleSlice.actions
+  setSelectedSubtitleId, setFocusSegmentTriggered, setFocusSegmentId, setFocusSegmentTriggered2,
+  setFocusToSegmentAboveId, setFocusToSegmentBelowId, setAspectRatio, setHasChanges } = subtitleSlice.actions
 
 // Export Selectors
 export const selectIsDisplayEditView = (state: RootState) =>
@@ -201,16 +209,18 @@ export const selectFocusSegmentTriggered2 = (state: { subtitleState: { focusSegm
   state.subtitleState.focusSegmentTriggered2
 // Hardcoding this value to achieve a desired size for the video player
 // TODO: Don't hardcode this value, instead make the video player component more flexible
-export const selectAspectRatio = (state: { subtitleState: { aspectRatios: subtitle["aspectRatios"] } }) =>
+export const selectAspectRatio = (_state: { subtitleState: { aspectRatios: subtitle["aspectRatios"] } }) =>
   50
 
 export const selectSubtitles = (state: { subtitleState: { subtitles: subtitle["subtitles"] } }) =>
   state.subtitleState.subtitles
-export const selectSelectedSubtitleFlavor = (state: { subtitleState: { selectedSubtitleFlavor: subtitle["selectedSubtitleFlavor"] } }) =>
-  state.subtitleState.selectedSubtitleFlavor
-export const selectSelectedSubtitleByFlavor = (state: { subtitleState:
-  { subtitles: subtitle["subtitles"]; selectedSubtitleFlavor: subtitle["selectedSubtitleFlavor"]; }; }) =>
-  state.subtitleState.subtitles[state.subtitleState.selectedSubtitleFlavor]
+export const selectSelectedSubtitleId = (state: { subtitleState: { selectedSubtitleId: subtitle["selectedSubtitleId"] } }) =>
+  state.subtitleState.selectedSubtitleId
+export const selectSelectedSubtitleById = (state: { subtitleState:
+  { subtitles: subtitle["subtitles"]; selectedSubtitleId: subtitle["selectedSubtitleId"]; }; }) =>
+  state.subtitleState.subtitles[state.subtitleState.selectedSubtitleId]
+export const selectHasChanges = (state: { subtitleState: { hasChanges: subtitle["hasChanges"] } }) =>
+  state.subtitleState.hasChanges
 
 /**
  * Alternative middleware to setCurrentlyAt.
