@@ -33,18 +33,18 @@ const workflowPostAndProcessSlice = createSlice({
   },
   extraReducers: builder => {
     builder.addCase(
-      postVideoInformationWithWorkflow.pending, (state, action) => {
+      postVideoInformationWithWorkflow.pending, (state, _action) => {
         state.status = 'loading'
-    })
+      })
     builder.addCase(
-      postVideoInformationWithWorkflow.fulfilled, (state, action) => {
+      postVideoInformationWithWorkflow.fulfilled, (state, _action) => {
         state.status = 'success'
-    })
+      })
     builder.addCase(
       postVideoInformationWithWorkflow.rejected, (state, action) => {
         state.status = 'failed'
         state.error = action.error.message
-    })
+      })
   }
 })
 
