@@ -28,24 +28,24 @@ const workflowPostSlice = createSlice({
   name: 'workflowPostState',
   initialState,
   reducers: {
-    resetPostRequestState: (state) => {
+    resetPostRequestState: state => {
       state.status = 'idle'
     }
   },
   extraReducers: builder => {
     builder.addCase(
-      postVideoInformation.pending, (state, action) => {
+      postVideoInformation.pending, (state, _action) => {
         state.status = 'loading'
-    })
+      })
     builder.addCase(
-      postVideoInformation.fulfilled, (state, action) => {
+      postVideoInformation.fulfilled, (state, _action) => {
         state.status = 'success'
-    })
+      })
     builder.addCase(
       postVideoInformation.rejected, (state, action) => {
         state.status = 'failed'
         state.error = action.error.message
-    })
+      })
   }
 })
 
@@ -59,7 +59,7 @@ interface segmentAPI {
 // Convert a segment from how it is stored in redux into
 // a segment that can be send to Opencast
 export const convertSegments = (segments: Segment[]) => {
-  let newSegments: segmentAPI[] = []
+  const newSegments: segmentAPI[] = []
 
   segments.forEach(segment => {
     newSegments.push({
