@@ -3,24 +3,24 @@ import { WebVTTParser, WebVTTSerializer } from 'webvtt-parser';
 import { ExtendedSubtitleCue, SubtitleCue } from '../types';
 
 export const roundToDecimalPlace = (num: number, decimalPlace: number) => {
-  let decimalFactor = Math.pow(10, decimalPlace)
+  const decimalFactor = Math.pow(10, decimalPlace)
   return Math.round((num + Number.EPSILON) * decimalFactor) / decimalFactor
 }
 
 
 // Returns a promise that resolves after `ms` milliseconds.
-export const sleep = (ms: number) => new Promise((resolve, reject) => setTimeout(resolve, ms));
+export const sleep = (ms: number) => new Promise((resolve, _reject) => setTimeout(resolve, ms));
 
 
 // Get an understandable time string for ARIA
 export const convertMsToReadableString = (ms: number): string => {
-  let hours = new Date((ms ? ms : 0)).toISOString().substr(11, 2)
-  let minutes = new Date((ms ? ms : 0)).toISOString().substr(14, 2)
-  let seconds = new Date((ms ? ms : 0)).toISOString().substr(17, 2)
+  const hours = new Date((ms ? ms : 0)).toISOString().substr(11, 2)
+  const minutes = new Date((ms ? ms : 0)).toISOString().substr(14, 2)
+  const seconds = new Date((ms ? ms : 0)).toISOString().substr(17, 2)
 
-  let result = []
-  if (parseInt(hours) > 0) { result.push(hours + " hours, ")}
-  if (parseInt(minutes) > 0 || parseInt(hours) > 0) { result.push(minutes + " minutes, ")}
+  const result = []
+  if (parseInt(hours) > 0) { result.push(hours + " hours, ") }
+  if (parseInt(minutes) > 0 || parseInt(hours) > 0) { result.push(minutes + " minutes, ") }
   result.push(seconds + " seconds")
 
   return result.join("")
@@ -32,9 +32,9 @@ export const convertMsToReadableString = (ms: number): string => {
  */
 export function safeJsonParse(str : string) {
   try {
-      return [null, JSON.parse(str)];
+    return [null, JSON.parse(str)];
   } catch (err) {
-      return [err];
+    return [err];
   }
 }
 
@@ -43,36 +43,36 @@ export function safeJsonParse(str : string) {
  * Currently, this cannot be checked with "@support", so we use this workaround
  * instead.
  */
-var flexGapIsSupported: boolean | undefined;
+let flexGapIsSupported: boolean | undefined;
 export function checkFlexGapSupport() {
   // Use the cached value if it has been defined
-	if (flexGapIsSupported !== undefined) {
-		return flexGapIsSupported
-	}
+  if (flexGapIsSupported !== undefined) {
+    return flexGapIsSupported
+  }
 
-	// Create a flex container with row-gap set
-	const flex = document.createElement('div')
-	flex.style.display = 'flex'
-	flex.style.flexDirection = 'column'
-	flex.style.rowGap = '1px'
-	flex.style.position = 'absolute'
+  // Create a flex container with row-gap set
+  const flex = document.createElement('div')
+  flex.style.display = 'flex'
+  flex.style.flexDirection = 'column'
+  flex.style.rowGap = '1px'
+  flex.style.position = 'absolute'
 
-	// Create two, elements inside it
-	flex.appendChild(document.createElement('div'))
-	flex.appendChild(document.createElement('div'))
+  // Create two, elements inside it
+  flex.appendChild(document.createElement('div'))
+  flex.appendChild(document.createElement('div'))
 
-	// Append to the DOM (needed to obtain scrollHeight)
-	document.body.appendChild(flex)
+  // Append to the DOM (needed to obtain scrollHeight)
+  document.body.appendChild(flex)
 
   // Flex container should be 1px high due to the row-gap
   flexGapIsSupported = flex.scrollHeight === 1
 
   // Remove element from the DOM after you are done with it
-  if(flex.parentNode) {
+  if (flex.parentNode) {
     flex.parentNode.removeChild(flex)
   }
 
-	return flexGapIsSupported
+  return flexGapIsSupported
 }
 
 /**
@@ -117,7 +117,7 @@ export function serializeSubtitle(subtitle: SubtitleCue[]) {
   return seri.serialize(cues)
 }
 
-export function parseSubtitle(subtitle: String) {
+export function parseSubtitle(subtitle: string): SubtitleCue[] {
   // Used parsing library: https://www.npmjs.com/package/webvtt-parser
   // - Unmaintained and does have bugs, so we will need to switch eventually
   // Other interesting vtt parsing libraries:
@@ -144,7 +144,7 @@ export function parseSubtitle(subtitle: String) {
   // Attach a unique id to each segment/cue
   // This is used by React to keep track of cues between changes (e.g. addition, deletion)
   let index = 0
-  for (let cue of tree.cues) {
+  for (const cue of tree.cues) {
     if (!cue.id) {
       cue.idInternal = nanoid()
       tree.cues[index] = cue
@@ -159,4 +159,24 @@ export function parseSubtitle(subtitle: String) {
   }
 
   return tree.cues
+}
+
+/**
+ * Parse language code to language name
+ * Returns language name in the language set in the browser
+ * Returns undefined if the input was undefined or the language code could not
+ * be parsed
+ */
+export function languageCodeToName(lang: string | undefined): string | undefined {
+  if (!lang) {
+    return undefined
+  }
+  const browserLang = window.navigator.language;
+  const languageNames = new Intl.DisplayNames(browserLang, { type: "language" });
+  try {
+    return languageNames.of(lang.trim());
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  } catch (e) {
+    return undefined
+  }
 }
